% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_caption_attribute.R
\name{add_caption_attribute}
\alias{add_caption_attribute}
\title{Adds caption attribute}
\usage{
add_caption_attribute(
  main_question,
  data_out,
  by_pos,
  tailored_group = NULL,
  translations = .saros.env$defaults$translations
)
}
\arguments{
\item{main_question}{String from get_main_question2}

\item{data_out}{Output from summarize_data}

\item{by_pos}{Named integer for the by-variable.}

\item{tailored_group}{String, indicating the name of the tailored group}

\item{translations}{List of named strings for by and (N=, etc. see getOption("saros")$translations.}
}
\value{
String
}
\description{
Adds caption attribute
}
