% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{add_category_order}
\alias{add_category_order}
\title{Add response category ordering (only useful for long format cat-cat tables)}
\usage{
add_category_order(data, sort_by = NULL)
}
\arguments{
\item{data}{Dataset}

\item{sort_by}{Sorting method for response categories}
}
\value{
Dataset with .category_order column added
}
\description{
Add response category ordering (only useful for long format cat-cat tables)
}
\keyword{internal}
