% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{add_indep_order}
\alias{add_indep_order}
\title{Add independent variable category ordering}
\usage{
add_indep_order(data, sort_by = ".factor_order", descend = FALSE)
}
\arguments{
\item{data}{Dataset}

\item{sort_by}{Sorting method for independent categories (NULL = no sorting)}

\item{descend}{Whether to reverse the order}
}
\value{
Dataset with .indep_order column added
}
\description{
Add independent variable category ordering
}
\keyword{internal}
