% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{calculate_sum_value_order}
\alias{calculate_sum_value_order}
\title{Calculate ordering based on .sum_value (for category-based sorting)}
\usage{
calculate_sum_value_order(data, descend = FALSE)
}
\arguments{
\item{data}{Dataset with .sum_value column}

\item{descend}{Logical indicating if sorting should be descending}
}
\value{
Numeric vector of ordering values
}
\description{
Calculate ordering based on .sum_value (for category-based sorting)
}
\keyword{internal}
