% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{center_string}
\alias{center_string}
\title{Center String Vector}
\usage{
center_string(string, maxwidth = 50)
}
\arguments{
\item{string}{String vector}

\item{maxwidth}{Maximum width}
}
\value{
String vector
}
\description{
Center String Vector
}
\examples{
center_string(string=c("This is a very long label for a graph.",
"But this one is even longer due to superfluous and verbose way of writing"),
 maxwidth=20)
}
