% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_categorical_office.R
\name{chart_categorical_office}
\alias{chart_categorical_office}
\title{Create Categorical Data Chart from Summarized Data}
\usage{
chart_categorical_office(
  data,
  ...,
  label_font_size = 10,
  main_font_size = 8,
  font_family = "Calibri",
  colour_palette = NULL,
  colour_na = "gray90",
  colour_2nd_binary_cat = "#ffffff",
  vertical = FALSE,
  percentage = TRUE,
  digits = 1,
  seed = 1,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{...}{Optional parameters forwarded from above.}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text
Must contain at least the number of unique values (excl. missing) in the data set.}

\item{font_family}{Word font family. See officer::fp_text}

\item{colour_palette}{[\code{character()}]\cr
Must contain at least the number of unique values (incl. missing) in the data set.}

\item{colour_na}{[\code{character(1)}]\cr Colour as a single string.}

\item{colour_2nd_binary_cat}{[\code{character(1)}]\cr Colour for second category in binary variables. Often useful to hide this.}

\item{vertical}{[\code{logical(1)}] Logical. If FALSE (default), then horizontal.}

\item{percentage}{[\code{logical(1)}] If TRUE, will compute percentage and add percentage sign to data label.}

\item{digits}{[\code{integer(1)}]\cr Number of digits in percentages.}

\item{seed}{[\code{integer(1)}] Optional random seed for selection of colours in blender.}

\item{call}{For internal use}
}
\value{
mschart-object. Can be added to an rdocx, rpptx or rxlsx object.
}
\description{
Create Categorical Data Chart from Summarized Data
}
\examples{
#chart_categorical_office(summarize_data(ex_survey1[paste0("b_", 1:3)]))
}
