% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cat_html.R
\name{chart_categorical_plot}
\alias{chart_categorical_plot}
\title{Create Single Interactive Categorical Plot with Univariates for Categorical Columns Sharing Same Categories}
\usage{
chart_categorical_plot(
  data,
  ...,
  label_font_size = 10,
  main_font_size = 8,
  font_family = "Calibri",
  colour_palette = NULL,
  colour_na = "gray90",
  colour_2nd_binary_cat = "#ffffff",
  vertical = FALSE,
  data_label = c("proportion", "percentage", "percentage_bare", "count", "mean",
    "median"),
  digits = 1,
  x_axis_label_width = 20,
  seed = 1,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{...}{Optional parameters forwarded from above.}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text
Must contain at least the number of unique values (excl. missing) in the data set.}

\item{font_family}{Word font family. See officer::fp_text}

\item{colour_palette}{[\code{character()}]\cr
Must contain at least the number of unique values (incl. missing) in the data set.}

\item{colour_na}{[\code{character(1)}]\cr Colour as a single string.}

\item{colour_2nd_binary_cat}{[\code{character(1)}]\cr Colour for second category in binary variables. Often useful to hide this.}

\item{vertical}{[\code{logical(1)}] Logical. If FALSE (default), then horizontal.}

\item{data_label}{[\code{character(1)}] What to display, one of "proportion", "percentage", "percentage_bare" (which is without the percentage symbol), "count", "mean", or "median". Defaults to "proportion".}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{x_axis_label_width}{[\code{integer(1)}] Width of the labels used for the categorical column names.}

\item{seed}{[\code{integer(1)}] Optional random seed for selection of colours in blender.}

\item{call}{For internal use}
}
\value{
ggiraph object, plottable with plot()
}
\description{
Create Single Interactive Categorical Plot with Univariates for Categorical Columns Sharing Same Categories
}
