% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combn_upto}
\alias{combn_upto}
\title{Create All Possible Combinations of Vector Elements with Minimum A and
Maximum B.}
\usage{
combn_upto(
  vec = c("a", "b", "c", "d", "e", "f", "g"),
  n_min = 6L,
  n_max = length(vec)
)
}
\arguments{
\item{vec}{Vector}

\item{n_min}{Minimum number of elements}

\item{n_max}{Maximum number of elements. Defaults to length of vec.}
}
\value{
A data frame
}
\description{
Create All Possible Combinations of Vector Elements with Minimum A and
Maximum B.
}
\examples{
combn_upto()
}
