% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_htaccess.R
\name{create__headers_file}
\alias{create__headers_file}
\title{Create a _headers file for Netlify publication}
\usage{
create__headers_file(
  site_path = "_site",
  mesos_paths = paste0("rapporter/Barnehageleder/2022H/mesos/", c("BI", "DMMH",
    "NHH_og_AFF", "OsloMet", "UiA", "UiT", "USN")),
  mesos_usernames = c("BI", "DMMH", "NHH_og_AFF", "OsloMet", "UiA", "UiT", "USN"),
  mesos_passwords = c("BI", "DMMH", "NHH_og_AFF", "OsloMet", "UiA", "UiT", "USN"),
  global_username = "admin",
  global_password = "arturead"
)
}
\arguments{
\item{site_path}{String, path to where to locate the _headers file}

\item{mesos_paths}{Character vector of relative paths within site_path to the locked directories}

\item{mesos_usernames, mesos_passwords}{Character vector of respective usernames and passwords for the mesos_paths}

\item{global_username, global_password}{String, username and password that always gives access to a mesos_paths path. Defaults to "admin"}
}
\value{
String, the path to the _headers file
}
\description{
Create a _headers file for Netlify publication
}
\examples{
create__headers_file(site_path=tempdir())
}
