% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_caption_attribute.R
\name{create_caption}
\alias{create_caption}
\title{Adds caption attribute}
\usage{
create_caption(
  main_question,
  data_out,
  indep_pos = NULL,
  mesos_group = NULL,
  filepath = NULL,
  translations = NULL
)
}
\arguments{
\item{main_question}{String from get_main_question2}

\item{data_out}{Output from summarize_data}

\item{indep_pos}{Named integer for the by-variable.}

\item{mesos_group}{String, indicating the name of the mesos group}

\item{filepath}{String, xlsx-file}

\item{translations}{List of named strings for by and (N=, etc. see getOption("saros")$translations.}
}
\value{
String
}
\description{
Adds caption attribute
}
