% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_htaccess.R
\name{create_htaccess}
\alias{create_htaccess}
\title{Create Bulk of .htaccess files in local folders, having valid AuthUserFile when uploaded to ftp-server at remote_basepath}
\usage{
create_htaccess(
  remote_basepath = "/home/",
  local_basepath = "reports/2023V/org",
  local_subfolders = basename(fs::dir_ls(local_basepath, all = FALSE, recurse = FALSE,
    type = "directory")),
  local_main_htpasswd_path = ".main_htpasswd_public",
  universal_usernames = c("admin", "nifu")
)
}
\arguments{
\item{remote_basepath}{String}

\item{local_basepath}{String}

\item{local_subfolders}{Character vector}

\item{local_main_htpasswd_path}{string. Path to main htpasswd file containing all usernames and passwords}

\item{universal_usernames}{Character vector. Usernames in local_main_htpasswd_path which always have access to folder}
}
\value{
paste0(local_folders, ".htaccess")
}
\description{
Create Bulk of .htaccess files in local folders, having valid AuthUserFile when uploaded to ftp-server at remote_basepath
}
\examples{
\dontrun{
create_htaccess(remote_basepath =
               "/home/stagiede/public_html/stephan/SSN/rapporter/2023V/org",
                local_basepath =
     "O - Outputs to be published/1 Main site - testing/rapporter/2023V/org",
                local_main_htpasswd_path =
 "C:/Users/py128/NIFU/Metode - Sensitivt - Sensitivt/.main_htpasswd_public",
universal_usernames = c("admin", "nifu"))
}
}
