% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{create_static_text_ordinal}
\alias{create_static_text_ordinal}
\title{Creates a structured list with text interpretations for a set of variables.}
\usage{
create_static_text_ordinal(
  data,
  cols = NULL,
  by = NULL,
  percentage = FALSE,
  contents = c("intro", "not_used_category", "mode_max", "value_max", "value_min",
    "value_diff", "mean_max", "mean_min", "mean_diff", "median_max", "median_min",
    "median_diff", "variance_max", "variance_min"),
  include_numbers = TRUE,
  sort_by = NULL,
  desc = FALSE,
  ignore_if_below = 0,
  n_top_bottom = 1,
  showNA = c("ifany", "always", "no"),
  label_separator = NULL,
  translations = getOption("saros")$translations,
  percent_sign = TRUE,
  digits = 1,
  require_common_categories = TRUE
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{percentage}{[\code{logical(1)}]\cr Whether to compute percentage or frequency (default).}

\item{contents}{The type of text interpretations to return, multiple allowed. Defaults to all.}

\item{include_numbers}{Whether or not to include the actual numbers (percentages, means) in parentheses.}

\item{sort_by}{[\code{character(1)}] Sort output (and collapse if requested). Defaults to none (NULL).
\itemize{
\item{".top"}{The percentage for the highest category available in the variable.}
\item{".upper"}{The sum of the percentages for the categories above the middle category.}
\item{".mid_upper"}{The sum of the percentages for the categories including and above the middle category.}
\item{".mid_lower"}{The sum of the percentages for the categories including and below the middle category.}
\item{".lower"}{The sum of the percentages for the categories below the middle category.}
\item{".bottom"}{The percentage for the lowest category available in the variable.}
\item{".variable_label"}{Sort by the variable labels.}
\item{".id"}{Sort by the variable names.}
\item{".by_group"}{The groups of the by argument.}
\item{[\code{character(1)}]}{Character vector of category labels to sum together.}
}}

\item{desc}{[\code{logical(1)}]\cr Reverse sorting of sort_by. Defaults to ascending order (FALSE).}

\item{ignore_if_below}{[\code{numeric(1)}] Whether to hide label if below this value.}

\item{n_top_bottom}{The number of top and bottom entries to report.}

\item{showNA}{[\code{logical(1)}]\cr Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{translations}{A list of translations for the template text. See getOption("saros").}

\item{percent_sign}{[\code{logical(1)}]\cr Whether to include percentage symbol on chart.}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{require_common_categories}{Whether to check if all questions share common categories.}
}
\value{
List
}
\description{
Creates a structured list with text interpretations for a set of variables.
}
\examples{
create_static_text_ordinal(ex_survey1,
                           cols = tidyselect::matches("e_"),
                           label_separator = " - ")

}
