% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable3.R
\name{crosstable3}
\alias{crosstable3}
\title{Internal function for fast cross-table}
\usage{
crosstable3(x, ...)
}
\arguments{
\item{x}{data.frame, survey object (requires \code{srvyr}-package)}

\item{...}{Dynamic dots}
}
\value{
Data.frame
}
\description{
Internal function for fast cross-table
}
