% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{crosstable_to_apa}
\alias{crosstable_to_apa}
\title{Title}
\usage{
crosstable_to_apa(
  x,
  label_separator = NULL,
  docx_template = NULL,
  caption_style = NULL,
  body_style = NULL,
  table_header_style = NULL,
  footer_style = NULL,
  caption_autonum = NULL,
  topcaption = TRUE,
  return_docx = TRUE
)
}
\arguments{
\item{x}{Crosstable}

\item{label_separator}{[\code{character(1)}]\cr If not NULL (default), will split labels into main- and sub-questions and create figure caption.}

\item{docx_template}{[\code{character(1) || officer::read_docx()}]\cr}

\item{caption_style}{[\code{character(1)}]\cr Word template style to be used for formatting chart. Defaults to "Normal".}

\item{body_style}{[\code{character(1)}]\cr Word style for table body cells.}

\item{table_header_style}{[\code{character(1)}]\cr Word style for table header cells.}

\item{footer_style}{[\code{character(1)}]\cr Word style for table footer.}

\item{caption_autonum}{Object obtained from \link[officer]{run_autonum}.}

\item{topcaption}{[\code{logical(1)}]\cr Place caption above (TRUE, default) or below table.}

\item{return_docx}{[\code{logical(1)}]\cr If FALSE, will return the table
rather than a rdocx object with the table inside it. Set to TRUE (default)
if piping multiple tables and charts together. Set to FALSE if you want to
continue modify the table.}
}
\value{
rdocx object, which can be saved with print() after loading the officer-package
}
\description{
Title
}
\examples{
library(officer)
crosstable_to_apa(x=crosstable::crosstable(ex_survey1,
cols = b_1:b_3, percent_pattern = "{p_col}"),
 label_separator=" - ")
}
