% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_post_render.R
\name{detect_ghostscript}
\alias{detect_ghostscript}
\title{Detect Ghostscript Binary on System PATH}
\usage{
detect_ghostscript()
}
\value{
String with the Ghostscript command name, or \code{NULL} if not found.
}
\description{
Detect Ghostscript Binary on System PATH
}
\keyword{internal}
