% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{determine_variable_basis}
\alias{determine_variable_basis}
\title{Determine variable column basis}
\usage{
determine_variable_basis(data_summary)
}
\arguments{
\item{data_summary}{Data frame with variable information}
}
\value{
String indicating column to use as basis
}
\description{
Consistent logic for determining whether to use .variable_label or .variable_name
}
\keyword{internal}
