% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest.R
\name{embed_bi_sigtest}
\alias{embed_bi_sigtest}
\title{Title}
\usage{
embed_bi_sigtest(
  data,
  cols,
  by,
  ...,
  col_type = NULL,
  by_type = NULL,
  reps = 1000,
  digits = 1,
  hide_test_if_n_below = 10,
  label_separator = NULL,
  return_raw = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{...}{Optional parameters forwarded from above.}

\item{col_type, by_type}{Optional string specifying data type ("int", "cat", "character", "factor", "numeric", "integer"). Can be obtained from data lookup.}

\item{reps}{Integer, number of permutations.}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{hide_test_if_n_below}{Integer, if N is below this value, pvalue will not be shown.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{return_raw}{[\code{logical(1)}] Whether to return the raw static chart. Defaults to FALSE.}

\item{call}{Error call function, usually not needed.}
}
\value{
Data frame
}
\description{
Title
}
\examples{
embed_bi_sigtest(data=ex_survey1, cols = a_1:a_9, by = x1_sex)
embed_bi_sigtest(data=ex_survey1, cols = b_1:b_3, by = x1_sex, label_separator=" - ")
embed_bi_sigtest(data=ex_survey1, cols = c_1:c_2, by = x1_sex)
}
