% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_freq_plot_html.R
\name{embed_cat_freq_plot}
\alias{embed_cat_freq_plot}
\title{Embed Interactive Categorical Plot}
\usage{
embed_cat_freq_plot(
  data,
  ...,
  cols = tidyselect::everything(),
  by = NULL,
  summarized_data = NULL,
  label_separator = NULL,
  tailored_group = NULL,
  translations = .saros.env$defaults$translations,
  html_interactive = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Optional parameters forwarded from above.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{summarized_data}{Currently not in use}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{tailored_group}{String, indicating name of tailored group.}

\item{translations}{List of named strings for by and (N=, etc. see getOption("saros")$translations.}

\item{html_interactive}{Flag, defaults to TRUE}

\item{call}{Error call function, usually not needed.}
}
\value{
ggplot
}
\description{
Embed Interactive Categorical Plot
}
\examples{
\dontrun{
embed_cat_freq_plot(data = ex_survey1, cols = b_1:b_3)
}
}
