% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_freq_plot_docx.R
\name{embed_cat_freq_plot_docx}
\alias{embed_cat_freq_plot_docx}
\title{Create Word Report with Univariates for Categorical Columns Sharing Same Categories}
\usage{
embed_cat_freq_plot_docx(
  data,
  ...,
  cols = tidyselect::everything(),
  by = NULL,
  summarized_data = NULL,
  label_separator = NULL,
  translations = getOption("saros")$translations
)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Optional parameters forwarded from above.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{translations}{List of named strings for by and (N=, etc. see getOption("saros")$translations.}

\item{docx_template}{[\code{character(1) || officer::read_docx()}]\cr
Either a filepath to a template file, or a rdocx-object from \link[officer]{read_docx}.}

\item{chart_formatting}{[\code{character(1)}]\cr Template style to be used for formatting chart}

\item{caption_style}{[\code{character(1)}]\cr Template style to be used for formatting chart. Defaults to "Normal".}

\item{caption_autonum}{Object obtained from \link[officer]{run_autonum}.}
}
\value{
rdocx object, which can be saved with print() after loading the officer-package
}
\description{
Create Word Report with Univariates for Categorical Columns Sharing Same Categories
}
\examples{
library(officer) # To save the rdocx object to disk
filepath <-
 ex_survey1 |>
 embed_cat_freq_plot_docx(cols = a_1:a_9, return_raw = FALSE) |>
 print(target = "test_docx_a19.docx")
file.remove(filepath)



 test_docx_b13 <-
   ex_survey1 |>
   embed_cat_freq_plot_docx(cols = b_1:b_3,
                       plot_height_multiplier = .3,
                       plot_height_fixed_constant = 1,
                       return_raw = FALSE)
\dontrun{
print(test_docx_b13, target = "test_docx_b13.docx")
file.remove("test_docx_b13.docx")
}
}
