% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{embed_cat_prop_plot}
\alias{embed_cat_prop_plot}
\title{Embed Interactive Categorical Plot (DEPRECATED!)}
\usage{
embed_cat_prop_plot(
  data,
  ...,
  dep = tidyselect::everything(),
  indep = NULL,
  colour_palette = NULL,
  mesos_group = NULL,
  html_interactive = TRUE,
  inverse = FALSE
)
}
\arguments{
\item{data}{\code{data.frame}, \code{tibble} or potentially a \code{srvyr}-object.}

\item{...}{Dynamic dots, arguments forwarded to underlying function(s).}

\item{dep}{\code{tidyselect}-syntax for dependent variable(s).}

\item{indep}{\code{tidyselect}-syntax for an optional independent variable.}

\item{colour_palette}{Character vector. Avoid using this.}

\item{mesos_group}{String}

\item{html_interactive}{Flag, whether to include interactivity.}

\item{inverse}{Flag, whether to flip plot or table.}
}
\description{
This function has been deprecated.
Use instead \code{\link[=makeme]{makeme()}}
}
