% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table.R
\name{embed_cat_table}
\alias{embed_cat_table}
\title{Embed Reactable Table}
\usage{
embed_cat_table(
  data,
  ...,
  cols = tidyselect::everything(),
  by = NULL,
  summarized_data = NULL,
  label_separator = NULL,
  tailored_group = NULL,
  translations = .saros.env$defaults$translations,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Further arguments passed down.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{tailored_group}{String, indicating the name of the tailored group}

\item{translations}{List of named strings for by and (N=, etc. see getOption("saros")$translations.}

\item{call}{Error call function, usually not needed.}

\item{information}{Which pre-computed information for each variable-category to display.}
}
\value{
A reactable object. If return_raw=FALSE, then just the table.
}
\description{
Embed Reactable Table
}
