% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cat_docx.R
\name{embed_cat_table_docx}
\alias{embed_cat_table_docx}
\title{Embed Table into Docx}
\usage{
embed_cat_table_docx(
  data,
  ...,
  cols = tidyselect::everything(),
  by = NULL,
  showNA = c("ifany", "always", "never"),
  data_label = c("proportion", "percentage", "percentage_bare", "count", "mean",
    "median"),
  digits = if (data_label == "proportion") 2 else 1,
  sort_by = NULL,
  vertical = FALSE,
  descend = FALSE,
  ignore_if_below = 0,
  label_separator = NULL,
  docx_template = NULL,
  body_style = NULL,
  table_header_style = NULL,
  footer_style = NULL,
  caption_autonum = NULL,
  topcaption = TRUE,
  return_raw = TRUE
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{...}{Optional parameters forwarded from above.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{showNA}{[\code{logical(1)}]\cr Whether to show NA in categorical variables (one of c("ifany", "always", "never"), like in table()).}

\item{data_label}{[\code{character(1)}] What to display, one of "proportion", "percentage", "percentage_bare" (which is without the percentage symbol), "count", "mean", or "median". Defaults to "proportion".}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{sort_by}{[\code{character(1)}] Sort output (and collapse if requested). Defaults to none (NULL).
\itemize{
\item{".top"}{The proportion for the highest category available in the variable.}
\item{".upper"}{The sum of the proportions for the categories above the middle category.}
\item{".mid_upper"}{The sum of the proportions for the categories including and above the middle category.}
\item{".mid_lower"}{The sum of the proportions for the categories including and below the middle category.}
\item{".lower"}{The sum of the proportions for the categories below the middle category.}
\item{".bottom"}{The proportions for the lowest category available in the variable.}
\item{".variable_label"}{Sort by the variable labels.}
\item{".id"}{Sort by the variable names.}
\item{".by_group"}{The groups of the by argument.}
\item{[\code{character(1)}]}{Character vector of category labels to sum together.}
}}

\item{vertical}{[\code{logical(1)}] Logical. If FALSE (default), then horizontal.}

\item{descend}{[\code{logical(1)}]\cr Reverse sorting of sort_by. Defaults to ascending order (FALSE).}

\item{ignore_if_below}{[\code{numeric(1)}] Whether to hide label if below this value.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{docx_template}{[\code{character(1) || officer::read_docx()}]\cr
Either a filepath to a template file, or a rdocx-object from \link[officer]{read_docx}.}

\item{body_style}{[\code{character(1)}]\cr Word style for table body cells.}

\item{table_header_style}{[\code{character(1)}]\cr Word style for table header cells.}

\item{footer_style}{[\code{character(1)}]\cr Word style for table footer.
rather than a rdocx object with the table inside it. Set to TRUE (default)
if piping multiple tables and charts together. Set to FALSE if you want to
continue modify the table.}

\item{caption_autonum}{Object obtained from \link[officer]{run_autonum}.}

\item{topcaption}{[\code{logical(1)}]\cr Place caption above (TRUE, default) or below table.}

\item{return_raw}{[\code{logical(1)}] Whether to return the raw static chart. Defaults to FALSE.}
}
\value{
rdocx object, which can be saved with print() after loading the officer-package
}
\description{
Embed Table into Docx
}
\examples{
library(officer)
embed_cat_table_docx(data = ex_survey1,
                     cols = b_1:b_3,
                     label_separator = " - ")
}
