% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table_html.R
\name{embed_cat_table_html}
\alias{embed_cat_table_html}
\title{Embed Reactable Table}
\usage{
embed_cat_table_html(
  data,
  ...,
  cols = tidyselect::everything(),
  by = NULL,
  showNA = c("ifany", "always", "never"),
  data_label = c("proportion", "percentage", "percentage_bare", "count", "mean",
    "median"),
  digits = 1,
  sort_by = NULL,
  vertical = FALSE,
  descend = FALSE,
  ignore_if_below = 0,
  information = c(".variable_label", ".category", ".count", ".count_se", ".proportion",
    ".proportion_se", ".mean", ".mean_se", ".mean_base", ".data_label",
    ".comb_categories", ".sum_value"),
  label_separator = NULL,
  return_raw = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{...}{Further arguments passed down.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{showNA}{[\code{logical(1)}]\cr Whether to show NA in categorical variables (one of c("ifany", "always", "never"), like in table()).}

\item{data_label}{[\code{character(1)}] What to display, one of "proportion", "percentage", "percentage_bare" (which is without the percentage symbol), "count", "mean", or "median". Defaults to "proportion".}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{sort_by}{[\code{character(1)}] Sort output (and collapse if requested). Defaults to none (NULL).
\itemize{
\item{".top"}{The proportion for the highest category available in the variable.}
\item{".upper"}{The sum of the proportions for the categories above the middle category.}
\item{".mid_upper"}{The sum of the proportions for the categories including and above the middle category.}
\item{".mid_lower"}{The sum of the proportions for the categories including and below the middle category.}
\item{".lower"}{The sum of the proportions for the categories below the middle category.}
\item{".bottom"}{The proportions for the lowest category available in the variable.}
\item{".variable_label"}{Sort by the variable labels.}
\item{".id"}{Sort by the variable names.}
\item{".by_group"}{The groups of the by argument.}
\item{[\code{character(1)}]}{Character vector of category labels to sum together.}
}}

\item{vertical}{[\code{logical(1)}] Logical. If FALSE (default), then horizontal.}

\item{descend}{[\code{logical(1)}]\cr Reverse sorting of sort_by. Defaults to ascending order (FALSE).}

\item{ignore_if_below}{[\code{numeric(1)}] Whether to hide label if below this value.}

\item{information}{Which pre-computed information for each variable-category to display.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{return_raw}{[\code{logical(1)}] Whether to return the raw static chart. Defaults to FALSE.}

\item{call}{Error call function, usually not needed.}
}
\value{
A reactable object. If return_raw=FALSE, then just the table.
}
\description{
Embed Reactable Table
}
\examples{
embed_cat_table_html(data=ex_survey1, cols = a_1:a_9)
embed_cat_table_html(data=ex_survey1, cols = a_1:a_9, by = x1_sex)
}
