% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_freq_plot_html.R
\name{embed_checkbox_freq_plot}
\alias{embed_checkbox_freq_plot}
\title{Embed Interactive Categorical Plot}
\usage{
embed_checkbox_freq_plot(
  data,
  ...,
  dep = tidyselect::everything(),
  indep = NULL,
  mesos_group = NULL,
  colour_palette = NULL,
  inverse = FALSE,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{\emph{Survey data}

\verb{obj:<data.frame>|obj:<tbl_df>} // Required

A data frame (or a srvyr-object) with the columns specified in the
chapter_overview 'dep_cat', etc columns.}

\item{...}{\emph{Dynamic dots}

<\href{https://rlang.r-lib.org/reference/dyn-dots.html}{\code{dynamic-dots}}>

Arguments forwarded to the corresponding functions that create the elements.}

\item{dep, indep}{\emph{Variable selections}

<\code{tidyselect}> // \emph{Default:} \code{NULL}, meaning everything for dep, nothing for indep.

Columns in \code{data}. Currently allows tidyselect-syntax, which will be removed.}

\item{mesos_group}{\emph{Specific group to compare with}

\verb{scalar<character>} // \emph{Default:} \code{NULL} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}

\item{colour_palette}{Character vector of colours.}

\item{inverse}{Flag, defaults to FALSE. If TRUE, swaps x-axis and faceting.}

\item{call}{\emph{Internal call}

\verb{obj:<call>} // \emph{Default:} \code{rlang::caller_env()} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}
}
\value{
ggplot
}
\description{
Embed Interactive Categorical Plot
}
\examples{
check_plot <- embed_checkbox_freq_plot(data = ex_survey1, dep = a_1:a_3)
}
