% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_table.R
\name{embed_chr_table_html}
\alias{embed_chr_table_html}
\title{Interactive table of text data (DEPRECATED)}
\usage{
embed_chr_table_html(data, dep, ..., mesos_group = NULL)
}
\arguments{
\item{data}{\code{data.frame}, \code{tibble} or potentially a \code{srvyr}-object.}

\item{dep}{\code{tidyselect}-syntax for dependent variable(s).}

\item{...}{Dynamic dots, arguments forwarded to underlying function(s).}

\item{mesos_group}{String}
}
\description{
This function has been deprecated.
Use instead \code{\link[saros.contents:makeme]{saros.contents::makeme()}}
}
