% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{embed_table_reactable}
\alias{embed_table_reactable}
\title{Embed Reactable Table}
\usage{
embed_table_reactable(
  data,
  cols = tidyselect::everything(),
  by = NULL,
  showNA = c("ifany", "always", "no"),
  percentage = TRUE,
  digits = 1,
  percent_sign = TRUE,
  sort_by = c(".variable_name", ".variable_label", ".category"),
  ignore_if_below = 1,
  information = c(".variable_label", ".category", ".count", ".count_se", ".proportion",
    ".proportion_se", ".mean", ".mean_se", ".mean_base", ".data_label",
    ".comb_categories", ".sum_value"),
  label_separator = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{showNA}{[\code{logical(1)}]\cr Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).}

\item{percentage}{[\code{logical(1)}]\cr Whether to compute percentage or frequency (default).}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{percent_sign}{[\code{logical(1)}]\cr Whether to include percentage symbol on chart.}

\item{sort_by}{[\code{character(1)}] Sort output (and collapse if requested). Defaults to none (NULL).
\itemize{
\item{".top"}{The percentage for the highest category available in the variable.}
\item{".upper"}{The sum of the percentages for the categories above the middle category.}
\item{".mid_upper"}{The sum of the percentages for the categories including and above the middle category.}
\item{".mid_lower"}{The sum of the percentages for the categories including and below the middle category.}
\item{".lower"}{The sum of the percentages for the categories below the middle category.}
\item{".bottom"}{The percentage for the lowest category available in the variable.}
\item{".variable_label"}{Sort by the variable labels.}
\item{".id"}{Sort by the variable names.}
\item{".by_group"}{The groups of the by argument.}
\item{[\code{character(1)}]}{Character vector of category labels to sum together.}
}}

\item{ignore_if_below}{[\code{numeric(1)}] Whether to hide label if below this value.}

\item{information}{Which pre-computed information for each variable-category to display.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{...}{Further arguments passed to reactable()}
}
\description{
Embed Reactable Table
}
