% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest.R
\name{embed_uni_sigtest}
\alias{embed_uni_sigtest}
\title{Title}
\usage{
embed_uni_sigtest(
  data,
  cols,
  ...,
  by = NULL,
  col_type = NULL,
  by_type = NULL,
  label_separator = NULL,
  translations = .saros.env$defaults$translations,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{...}{Optional parameters forwarded from above.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{col_type, by_type}{Optional string specifying data type ("int", "cat", "character", "factor", "numeric", "integer"). Can be obtained from data lookup.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{translations}{List of named strings for by and (N=, etc. see getOption("saros")$translations.}

\item{call}{Error call function, usually not needed.}
}
\value{
Data frame
}
\description{
Title
}
\examples{
embed_uni_sigtest(data=ex_survey1, cols = a_1:a_9)
embed_uni_sigtest(data=ex_survey1, cols = b_1:b_3, label_separator=" - ")
embed_uni_sigtest(data=ex_survey1, cols = c_1:c_2)
}
