% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_post_render.R
\name{escape_for_regex}
\alias{escape_for_regex}
\title{Escape a String for Literal Use in a Regular Expression}
\usage{
escape_for_regex(x)
}
\arguments{
\item{x}{Character string to escape.}
}
\value{
String with regex metacharacters escaped.
}
\description{
Escape a String for Literal Use in a Regular Expression
}
\keyword{internal}
