% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_survey_elements_list}
\alias{ex_survey_elements_list}
\title{ex_survey_elements_list: List of elements (currently only uni_cat_prop_plot_html) for the ex_survey1}
\format{
A nested list of length one, where the unitary element consists
of multiple elements, for each of the data_overview combinations:
}
\usage{
ex_survey_elements_list
}
\description{
ex_survey_elements_list: List of elements (currently only uni_cat_prop_plot_html) for the ex_survey1
}
\keyword{datasets}
