% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_file_links.R
\name{extract_pdf_title}
\alias{extract_pdf_title}
\title{Extract Title from PDF File}
\usage{
extract_pdf_title(filepath)
}
\arguments{
\item{filepath}{Path to PDF file}
}
\value{
Character string with title or NULL
}
\description{
Extract Title from PDF File
}
\keyword{internal}
