% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_file_links.R
\name{extract_pptx_title}
\alias{extract_pptx_title}
\title{Extract Title from PPTX File}
\usage{
extract_pptx_title(filepath)
}
\arguments{
\item{filepath}{Path to PPTX file}
}
\value{
Character string with title or NULL
}
\description{
Extract Title from PPTX File
}
\keyword{internal}
