% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_sanitizer.R
\name{filename_sanitizer}
\alias{filename_sanitizer}
\title{Filename sanitizer replacing space and punctuation with underscore}
\usage{
filename_sanitizer(x, max_chars = NA_integer_)
}
\arguments{
\item{x}{Character vector of filenames/foldernames}

\item{max_chars}{Maximum character length}
}
\value{
Character vector of same length as x
}
\description{
Filename sanitizer replacing space and punctuation with underscore
}
\examples{
filename_sanitizer(c("Too long a name", "with invalid *^/&#"))
}
