% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_element_and_qmd_snippet.R
\name{gen_element_and_qmd_snippet}
\alias{gen_element_and_qmd_snippet}
\title{Mass Create Elements of A Certain Type}
\usage{
gen_element_and_qmd_snippet(
  data_overview_section,
  element_name = "uni_cat_prop_plot",
  data,
  summarized_data = NULL,
  element_folderpath_absolute,
  element_folderpath_relative,
  grouping_structure = NULL,
  translations = .saros.env$defaults$translations,
  ...,
  call = rlang::caller_env()
)
}
\arguments{
\item{data_overview_section}{A (grouped) data frame containing at least the following columns
\describe{
\item{<grouping variables>}{ used to distinguish sets of columns belonging to each element.}
\item{"col_name"}{ name of columns.}
\item{"designated_type"}{ either 'cat' (categorical/ordinal/binary), 'int' (interval/continuous) or 'chr' (text).}
}}

\item{element_name}{See \code{element_list()} for options.}

\item{data}{Survey data.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> arguments forwarded to the corresponding functions that create the elements.}
}
\value{
Named list of elements, where each element can .
}
\description{
Mass Create Elements of A Certain Type
}
