% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_qmd_chapters.R
\name{gen_qmd_chapters}
\alias{gen_qmd_chapters}
\title{Generate A Quarto Survey Report}
\usage{
gen_qmd_chapters(
  data_overview,
  data,
  element_names = NULL,
  label_separator = NULL,
  report_yaml_file = NULL,
  chapter_yaml_file = NULL,
  translations = .saros.env$defaults$translations,
  path = getwd(),
  ...,
  call = rlang::caller_env()
)
}
\arguments{
\item{data_overview}{A (possibly grouped) tbl with overview of the survey columns in data.}

\item{report_yaml_file, chapter_yaml_file}{Optional ymlthis-objects. Defaults will be obtained from getOptions("ymlthis.default_options") if NULL. The chapter_yaml_file will be used across all chapters. Authors will always be taken from the data_overview.}

\item{path}{String specifying the path to the working directory. Defaults to the current working directory.}

\item{elements}{A named list of elements used to populate each chapter's section. The names of the list must be one of \code{\link{list_available_element_types}}. Defaults to a list of TRUE values for various elements. Each named element of the list can contain
\describe{
\item{\code{NULL}}{(default)}{ no element is inserted.}
\item{\code{list()}}{A named sublist of already produced objects of the same type. The names must be of }
\item{\code{character()}}{A character vector of paths to Rds-files containing the objects. These will be read in as needed, saving memory.}
}}

\item{index_filename}{String specifying the name of the main index Quarto file (and its subfolder) used to collect all the chapters. Defaults to "index.qmd".}
}
\value{
A Quarto survey report generated in the specified working directory.
}
\description{
This function generates a Quarto survey report.
}
\details{
A report consists of multiple chapters and an index file that merges them together.
A chapter can contain any user-defined set of dependent, independent or bivariate variable sets.
A chapter consists of multiple sections.
A section is defined as a group in the data_overview (ignoring the chapter grouping level) containing variables of the same type, meaning at a minimum that the variables in the section sharing the same response options, the same main question, and being of the same data type.
}
