% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_qmd_index.R
\name{gen_qmd_index}
\alias{gen_qmd_index}
\title{Generate Quarto Index File That Merges All Chapters}
\usage{
gen_qmd_index(
  yaml_file = NULL,
  authors = NULL,
  index_filepath = "complete_report.qmd",
  chapter_filepaths = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{yaml_file}{A ymlthis-object. Defaults will be obtained from getOptions("ymlthis.default_options") if NULL.}

\item{authors}{A character vector listing the authors of the Quarto document. If not provided, the author information will be omitted.}

\item{index_filepath}{A string specifying the name of the Quarto index file to be created. Defaults to "complete_report.qmd".}

\item{chapter_filepaths}{A character vector containing the filepaths of the chapter files to be included in the index, in the desired order.}

\item{call}{Internal call argument. Not to be fiddled with by the user.}
}
\value{
A string containing the filepath to the generated Quarto index file.
}
\description{
This function creates an index Quarto file (QMD) that merges all chapters in the specified order. It can also include optional title and author(s) information.
}
\examples{
filepath <- tempfile(pattern = "index", fileext = ".qmd")
index_file <-
  saros:::gen_qmd_index(
                 authors = c("Marky Twain",
                             "J.K. Rawlings",
                             "Stephen Prince"),
               chapter_filepaths =
                 c("Ch1.qmd",
                 "Ch2.qmd",
                 "Ch3.qmd"),
                 index_filepath = filepath)
if(interactive()) browseURL(filepath)
unlink(filepath)
}
