% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_qmd_report.R
\name{gen_qmd_report}
\alias{gen_qmd_report}
\title{Generate A Quarto Survey Report}
\usage{
gen_qmd_report(
  data_overview,
  elements = NULL,
  glue_index_string = NULL,
  ignore_if_below = 0,
  captions = c("asis", "pretty", "none"),
  report_ymlthis_config = NULL,
  chapter_ymlthis_config = NULL,
  index_filename = "index.qmd",
  path = getwd()
)
}
\arguments{
\item{data_overview}{A (possibly grouped) tbl with overview of the survey columns in data.}

\item{elements}{A named list of elements used to populate each chapter's section. The names of the list must be one of \code{\link{list_available_element_types}}. Defaults to a list of TRUE values for various elements. Each named element of the list can contain
\describe{
\item{\code{NULL}}{(default)}{ no element is inserted.}
\item{\code{list()}}{A named sublist of already produced objects of the same type. The names must be of }
\item{\code{character()}}{A character vector of paths to Rds-files containing the objects. These will be read in as needed, saving memory.}
}}

\item{glue_index_string}{String used to glue together columns in the data_overview for creating an identifier to look up objects in elements-lists and character vectors. Defaults to "{designated_role}_{col_group}_{name_prefix}".}

\item{ignore_if_below}{A double between 0 and 1 to control which elements are hidden based on their significance level. Defaults to 0 (show all).}

\item{captions}{A string indicating whether the caption for each plot/table is displayed. One of
\describe{
\item{\code{"asis"}}{A technical ID, which is useful for troubleshooting or quick search and replace at the end.}
\item{\code{"pretty"}}{A full text title in English. Currently no translations have been implemented.}
\item{\code{"none"}}{None.}
}}

\item{report_ymlthis_config, chapter_ymlthis_config}{Optional ymlthis-objects. Defaults will be obtained from getOptions("ymlthis.default_options") if NULL. The chapter_ymlthis_config will be used across all chapters. Authors will always be taken from the data_overview.}

\item{index_filename}{String specifying the name of the main index Quarto file (and its subfolder) used to collect all the chapters. Defaults to "index.qmd".}

\item{path}{String specifying the path to the working directory. Defaults to the current working directory.}
}
\value{
A Quarto survey report generated in the specified working directory.
}
\description{
This function generates a Quarto survey report.
}
\details{
A report consists of multiple chapters and an index file that merges them together.
A chapter can contain any user-defined set of dependent, independent or bivariate variable sets.
A chapter consists of multiple sections.
A section is defined as a group in the data_overview (ignoring the chapter grouping level) containing variables of the same type, meaning at a minimum that the variables in the section sharing the same response options, the same main question, and being of the same data type.
}
\examples{
# Define temporary folder for storing the elements
library(dplyr)
temp_folder <- tempdir()
data_overview <-
 ex_survey_ch_overview \%>\%
  saros::refine_data_overview(data = ex_survey1,
                              label_separator = " - ",
                              name_separator = "_")

if(!is.null(quarto::quarto_path())) {
  index_filepath <-
   gen_qmd_report(
      data_overview = data_overview,
      elements = ex_survey_elements_list,
      captions = "pretty",
      path = temp_folder)

  if(interactive()) {
    browseURL(fs::path(temp_folder, "index.qmd"))
    quarto::quarto_render(fs::path(temp_folder, "index.qmd"))
    browseURL(fs::path(temp_folder, "index.html"))
  }
}
unlink(temp_folder)
}
