% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_display_column}
\alias{get_data_display_column}
\title{Determine display column based on data availability}
\usage{
get_data_display_column(data)
}
\arguments{
\item{data}{Data frame containing variable information}
}
\value{
Character string indicating which column to use
}
\description{
Checks if .variable_label column exists and has non-NA values to determine
whether to use .variable_label or .variable_name for display.
}
\keyword{internal}
