% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_data_label_opts}
\alias{get_data_label_opts}
\title{Get Valid Data Labels for Figures and Tables}
\usage{
get_data_label_opts()
}
\value{
Character vector
}
\description{
Get Valid Data Labels for Figures and Tables
}
