% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_element_path.R
\name{get_element_path}
\alias{get_element_path}
\title{Retrieve Element from Elements List}
\usage{
get_element_path(
  data_overview = NULL,
  elements = NULL,
  glue_index_string = NULL,
  ignore_if_below = 0,
  path,
  call = rlang::caller_env()
)
}
\arguments{
\item{data_overview}{A (possibly grouped) tbl with overview of the survey columns in data.}

\item{elements}{A named list of elements used to populate each chapter's section. The names of the list must be one of \code{\link{list_available_element_types}}. Defaults to a list of TRUE values for various elements. Each named element of the list can contain
\describe{
\item{\code{NULL}}{(default)}{ no element is inserted.}
\item{\code{list()}}{A named sublist of already produced objects of the same type. The names must be of }
\item{\code{character()}}{A character vector of paths to Rds-files containing the objects. These will be read in as needed, saving memory.}
}}

\item{glue_index_string}{String used to glue together columns in the data_overview for creating an identifier to look up objects in elements-lists and character vectors. Defaults to "{designated_role}_{col_group}_{name_prefix}".}

\item{ignore_if_below}{A double between 0 and 1 to control which elements are hidden based on their significance level. Defaults to 0 (show all).}

\item{path}{String specifying the path to the working directory. Defaults to the current working directory.}

\item{call}{Internal.}
}
\value{
Either NULL or a filepath to the object as a string. Returns NULL if element_contents is NULL.
}
\description{
Retrieve Element from Elements List
}
