% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{get_indep_col_name}
\alias{get_indep_col_name}
\title{Get the name of the independent variable column}
\usage{
get_indep_col_name(data)
}
\arguments{
\item{data}{Dataset}
}
\value{
Character string with column name, or NULL if not found
}
\description{
Get the name of the independent variable column
}
\keyword{internal}
