% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{get_indep_labels}
\alias{get_indep_labels}
\title{Get independent variable labels}
\usage{
get_indep_labels(dots)
}
\arguments{
\item{dots}{List from rlang::list2(...)}
}
\value{
Character vector of processed labels
}
\description{
Process independent variable labels with consistent logic across table functions
}
\keyword{internal}
