% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_labels.R
\name{get_raw_labels}
\alias{get_raw_labels}
\title{Helper function to extract raw variable labels from the data}
\usage{
get_raw_labels(data, col_pos = NULL, return_as_list = FALSE)
}
\arguments{
\item{data}{Dataset}

\item{col_pos}{Optional, character vector of column names or integer vector of positions}

\item{return_as_list}{Flag, whether to return as list or character vector}
}
\value{
List or character vector
}
\description{
Helper function to extract raw variable labels from the data
}
\keyword{internal}
