% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{get_standard_column_renamer}
\alias{get_standard_column_renamer}
\title{Get standard column renaming function}
\usage{
get_standard_column_renamer(
  main_question = "",
  use_header = FALSE,
  column_mappings = NULL
)
}
\arguments{
\item{main_question}{Main question for header}

\item{use_header}{Whether to use main question as header}

\item{column_mappings}{Named list of additional column mappings}
}
\value{
Function for renaming columns
}
\description{
Standardized column renaming logic for table functions
}
\keyword{internal}
