% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{get_target_categories}
\alias{get_target_categories}
\title{Get target categories for positional sorting}
\usage{
get_target_categories(data, method)
}
\arguments{
\item{data}{Dataset with .category column}

\item{method}{Positional method (.top, .bottom, .upper, .lower, etc.)}
}
\value{
Character vector of target category names
}
\description{
Uses subset_vector to determine which categories to include based on
positional methods like .top, .bottom, .upper, .lower, etc.
}
\keyword{internal}
