% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_settings.R
\name{global_settings_get}
\alias{global_settings_get}
\title{Get Global Options for saros-functions}
\usage{
global_settings_get(fn_name = "makeme")
}
\arguments{
\item{fn_name}{String, one of \code{"make_link"}, \code{"fig_height_h_barchart"} and \code{"makeme"}.}
}
\value{
List with options in R
}
\description{
Get Global Options for saros-functions
}
\examples{
global_settings_get()
}
