% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{improve_text}
\alias{improve_text}
\title{Improves the generated text by ChatGPT (v 14.03.23) magic.}
\usage{
improve_text(x, log = NULL, temperature = 0.5)
}
\arguments{
\item{x}{String}

\item{log}{Previous messages as outputted from the function.}

\item{temperature}{How creative the responses are, from 0 to 2, default in ChatGPT is 1, here set to .5 to avoid changes to the facts.}
}
\value{
List with response and input settings.
}
\description{
Improves the generated text by ChatGPT (v 14.03.23) magic.
}
