% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{is_colour}
\alias{is_colour}
\title{Are All Colours in Vector Valid Colours}
\usage{
is_colour(x)
}
\arguments{
\item{x}{Character vector of colours in hex-format.}
}
\value{
Logical, or error.
}
\description{
As title says. From: http://stackoverflow.com/a/13290832/3315962
}
\examples{
is_colour(c("#ff00ff", "#010101"))
}
