% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_labels.R
\name{is_string}
\alias{is_string}
\title{Is x A String?}
\usage{
is_string(x)
}
\arguments{
\item{x}{Object}
}
\value{
Logical value.
}
\description{
Returns TRUE if object is a character of length 1.
}
\keyword{internal}
