% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_saros_elements.R
\name{lst_saros_elements}
\alias{lst_saros_elements}
\title{Mass Create Elements of A Certain Type}
\usage{
lst_saros_elements(
  data_overview,
  element_name = "uni_cat_plot_html",
  data = NULL,
  ...
)
}
\arguments{
\item{data_overview}{A (grouped) data frame containing at least the following columns
\describe{
\item{<grouping variables>}{ used to distinguish sets of columns belonging to each element.}
\item{"col_name"}{ name of columns.}
\item{"designated_type"}{ either 'cat' (categorical/ordinal/binary), 'int' (interval/continuous) or 'chr' (text).}
}}

\item{element_name}{See \code{element_list()} for options.}

\item{data}{Survey data.}

\item{...}{arguments forwarded to the corresponding functions that create the elements.}
}
\value{
Named list of elements, where each element can .
}
\description{
Mass Create Elements of A Certain Type
}
