% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_content.R
\name{make_content}
\alias{make_content}
\title{Method for Creating Saros Contents}
\usage{
make_content(type, ...)
}
\arguments{
\item{type}{\emph{Method name}

\verb{scalar<character>} with a class named by itself.

Optional string indicating the specific method. Occasionally
useful for error messages, etc.}

\item{...}{\emph{Dots}

Arguments provided by \code{makeme}}
}
\value{
The returned object class depends on the type.
\code{type="*_table_html"} always returns a \code{tibble}.
\code{type="*_plot_html"} always returns a \code{ggplot}.
\code{type="*_docx"} always returns a \code{rdocx} object if \code{path=NULL},
or has side-effect of writing docx file to disk if \code{path} is set.
}
\description{
Takes the same arguments as \code{makeme}, except
that dep and indep in make_content are character vectors,
for ease of user-customized function programming.
}
