% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_link.R
\name{make_link.list}
\alias{make_link.list}
\title{make_link}
\usage{
\method{make_link}{list}(data, ..., separator_list_items = ". ")
}
\arguments{
\item{data}{\emph{Data or object}

\verb{<data.frame|tbl|obj>}

Data frame if using a tabular data \code{save_fn}, or possibly any
R object, if a serializing \code{save_fn} is provided (e.g. \code{\link[=saveRDS]{saveRDS()}}).}

\item{...}{\emph{Dynamic dots}

<\href{https://rlang.r-lib.org/reference/dyn-dots.html}{\code{dynamic-dots}}>

Arguments forwarded to the corresponding functions that create the elements.}

\item{separator_list_items}{\emph{Separator string between multiple list items}

\verb{scalar<character>} // \emph{default:} \code{". "} (\code{optional})}
}
\description{
make_link
}
