% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_link.R
\name{make_link}
\alias{make_link}
\title{Save data to a file and return a Markdown link}
\usage{
make_link(data, ...)
}
\arguments{
\item{data}{\emph{Data or object}

\verb{<data.frame|tbl|obj>}

Data frame if using a tabular data \code{save_fn}, or possibly any
R object, if a serializing \code{save_fn} is provided (e.g. \code{\link[=saveRDS]{saveRDS()}}).}

\item{...}{\emph{Dynamic dots}

<\href{https://rlang.r-lib.org/reference/dyn-dots.html}{\code{dynamic-dots}}>

Arguments forwarded to the corresponding functions that create the elements.}
}
\value{
String.
}
\description{
The file is automatically named by a hash of the object, removing the need
to come up with unique file names inside a Quarto report. This has the
added benefit of reducing storage needs if the objects needing linking to
are identical, and all are stored in the same folder. It also allows the user
to download multiple files without worrying about accidentally overwriting them.
}
\examples{
make_link(mtcars, folder = tempdir())
}
