% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_saros_elements.R
\name{mass_lst_saros_elements}
\alias{mass_lst_saros_elements}
\title{Mass Generate Elements}
\usage{
mass_lst_saros_elements(
  element_names = list_available_element_types(),
  data_overview,
  data,
  ...
)
}
\arguments{
\item{element_names}{Character vector of element names. Defaults to all available.}

\item{data_overview}{A (grouped) data frame containing at least the following columns
\describe{
\item{<grouping variables>}{ used to distinguish sets of columns belonging to each element.}
\item{"col_name"}{ name of columns.}
\item{"designated_type"}{ either 'cat' (categorical/ordinal/binary), 'int' (interval/continuous) or 'chr' (text).}
}}

\item{data}{Survey data.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> arguments forwarded to the corresponding functions that create the elements.}
}
\value{
List of elements that can be directly used in gen_qmd_report()
}
\description{
Mass Generate Elements
}
\examples{
library(dplyr)
ex_survey_elements_list <-
mass_lst_saros_elements(
               element_names = c("uni_cat_plot_html", "bi_catcat_plot_html"),
               data_overview =
                          ex_survey_ch_overview \%>\%
                          refine_data_overview(data = ex_survey1,
                                                      label_separator = " - ",
                                                      name_separator = "_"),
                        data = ex_survey1,
                        label_separator = " - ")
}
