% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_rng.R
\name{n_rng}
\alias{n_rng}
\title{Obtain range of N for a given data set and other settings.}
\usage{
n_rng(
  data,
  dep,
  indep = NULL,
  crowd = "all",
  mesos_var = NULL,
  mesos_group = NULL,
  glue_template_1 = "{n}",
  glue_template_2 = "[{n[1]}-{n[2]}]"
)
}
\arguments{
\item{data}{Dataset}

\item{dep, indep}{Character vector, names of (in)dependent variables}

\item{crowd}{String, one of "all", "target" or "others".}

\item{mesos_var}{Optional, NULL or string specifying name of variable used to
split dataset.}

\item{mesos_group}{Optional, NULL or string specifying value in \code{mesos_var}
indicating the target group.}

\item{glue_template_1, glue_template_2}{String, for the case of a single
value (1) or a range with minimum-maximum of values (2).}
}
\value{
Always a string.
}
\description{
Obtain range of N for a given data set and other settings.
}
\keyword{internal}
