% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_rng.R
\name{n_rng2}
\alias{n_rng2}
\title{Obtain range of N for a given \code{ggobj}.}
\usage{
n_rng2(ggobj, glue_template_1 = "{n}", glue_template_2 = "[{n[1]}-{n[2]}]")
}
\arguments{
\item{ggobj}{A \code{ggplot2}-object.}

\item{glue_template_1, glue_template_2}{String, for the case of a single
value (1) or a range with minimum-maximum of values (2).}
}
\value{
Always a string.
}
\description{
Obtain range of N for a given \code{ggobj}.
}
\keyword{internal}
