% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_replace_docx_imgs_with_mscharts.R
\name{post_render_docx_img_replacer}
\alias{post_render_docx_img_replacer}
\title{Post-Render All docx-files Found in The Output Folder To Replace Images with Mscharts}
\usage{
post_render_docx_img_replacer(path = Sys.getenv("QUARTO_PROJECT_OUTPUT_DIR"))
}
\arguments{
\item{path}{String, path to _site output folder}
}
\value{
Nothing, side-effects only.
}
\description{
Post-Render All docx-files Found in The Output Folder To Replace Images with Mscharts
}
