% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_freq_plot_docx.R
\name{prep_cat_freq_plot_docx}
\alias{prep_cat_freq_plot_docx}
\title{Create Categorical Data Chart from Summarized Data}
\usage{
prep_cat_freq_plot_docx(
  data,
  ...,
  label_font_size = 10,
  main_font_size = 8,
  font_family = "Calibri",
  colour_palette = NULL,
  colour_na = "gray90",
  vertical = FALSE,
  data_label = c("count", "proportion", "percentage", "percentage_bare", "mean",
    "median"),
  digits = if (data_label == "proportion") 2 else if (data_label == "count") 0 else 1,
  seed = 1,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{...}{Optional parameters forwarded from above.}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text
Must contain at least the number of unique values (excl. missing) in the data set.}

\item{font_family}{Word font family. See officer::fp_text}

\item{colour_palette}{[\code{character()}]\cr
Must contain at least the number of unique values (incl. missing) in the data set.}

\item{colour_na}{[\code{character(1)}]\cr Colour as a single string.}

\item{vertical}{[\code{logical(1)}] Logical. If FALSE (default), then horizontal.}

\item{data_label}{[\code{character(1)}] What to display, one of "proportion", "percentage", "percentage_bare" (which is without the percentage symbol), "count", "mean", or "median". Defaults to "proportion".}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{seed}{[\code{integer(1)}] Optional random seed for selection of colours in blender.}

\item{call}{For internal use}
}
\value{
mschart-object. Can be added to an rdocx, rpptx or rxlsx object.
}
\description{
Create Categorical Data Chart from Summarized Data
}
\examples{
saros:::prep_cat_freq_plot_docx(saros:::summarize_data(ex_survey1[paste0("b_", 1:3)]))
}
