% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_freq_plot_html.R
\name{prep_cat_freq_plot_html}
\alias{prep_cat_freq_plot_html}
\title{Create Single Interactive Categorical Plot with Univariates for Categorical Columns Sharing Same Categories}
\usage{
prep_cat_freq_plot_html(data, ..., call = rlang::caller_env())
}
\arguments{
\item{data}{Data frame.}

\item{...}{Optional parameters forwarded from above.}

\item{call}{Error call function, usually not needed.}
}
\value{
ggiraph object, plottable with plot()
}
\description{
Create Single Interactive Categorical Plot with Univariates for Categorical Columns Sharing Same Categories
}
