% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_prop_plot_html.R
\name{prep_cat_prop_plot_html}
\alias{prep_cat_prop_plot_html}
\title{Create Single Interactive Categorical Plot with Univariates for Categorical Columns Sharing Same Categories}
\usage{
prep_cat_prop_plot_html(data, ..., call = rlang::caller_env())
}
\arguments{
\item{data}{Data frame.}

\item{...}{Optional parameters forwarded from above.}

\item{call}{For internal use}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text
Must contain at least the number of unique values (excl. missing) in the data set.}

\item{colour_palette}{[\code{character()}]\cr
Must contain at least the number of unique values (incl. missing) in the data set.}

\item{colour_na}{[\code{character(1)}]\cr Colour as a single string.}

\item{colour_2nd_binary_cat}{[\code{character(1)}]\cr Colour for second category in binary variables. Often useful to hide this.}

\item{font_family}{Word font family. See officer::fp_text}

\item{vertical}{[\code{logical(1)}] Logical. If FALSE (default), then horizontal.}

\item{x_axis_label_width}{[\code{integer(1)}] Width of the labels used for the categorical column names.}

\item{seed}{[\code{integer(1)}] Optional random seed for selection of colours in blender.}
}
\value{
ggiraph object, plottable with plot()
}
\description{
Create Single Interactive Categorical Plot with Univariates for Categorical Columns Sharing Same Categories
}
