% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table_docx.R
\name{prep_cat_table_docx}
\alias{prep_cat_table_docx}
\title{Title}
\usage{
prep_cat_table_docx(
  data,
  body_style = NULL,
  table_header_style = NULL,
  footer_style = NULL
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{body_style}{[\code{character(1)}]\cr Word style for table body cells.}

\item{table_header_style}{[\code{character(1)}]\cr Word style for table header cells.}

\item{footer_style}{[\code{character(1)}]\cr Word style for table footer.
rather than a rdocx object with the table inside it. Set to TRUE (default)
if piping multiple tables and charts together. Set to FALSE if you want to
continue modify the table.}
}
\value{
rdocx object, which can be saved with print() after loading the officer-package
}
\description{
Title
}
