% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_chunk.R
\name{prepare_chunk}
\alias{prepare_chunk}
\title{Internal function to prepare a chunk for a Quarto report.}
\usage{
prepare_chunk(element_name, ...)
}
\arguments{
\item{element_name}{String, see draft_report}

\item{...}{Dynamic dots passed onto the methods.}
}
\value{
Returns a text string for the chunk, but
also has side-effects for generating files.
}
\description{
Internal function to prepare a chunk for a Quarto report.
}
