% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowd_plots_as_officer.R
\name{print.saros_officer_plots}
\alias{print.saros_officer_plots}
\title{Print method for saros_officer_plots}
\usage{
\method{print}{saros_officer_plots}(x, ...)
}
\arguments{
\item{x}{A saros_officer_plots object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print method for saros_officer_plots
}
\keyword{internal}
