% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{process_categorical_na}
\alias{process_categorical_na}
\title{Process categorical data for showNA settings}
\usage{
process_categorical_na(data, dots)
}
\arguments{
\item{data}{Data frame with .category column}

\item{dots}{List with showNA and indep settings}
}
\value{
Processed data frame
}
\description{
Handle NA categories based on showNA parameter for categorical tables
}
\keyword{internal}
