% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_draft_report_arguments_in_yaml.R
\name{read_default_draft_report_args}
\alias{read_default_draft_report_args}
\title{Read Default Arguments for \code{draft_report()} from YAML-file}
\usage{
read_default_draft_report_args(path = "settings.yaml")
}
\arguments{
\item{path}{\verb{scalar<character>} // Required. \emph{default:} \code{settings.yaml}}
}
\value{
The defaults as a \code{yaml}-object.
}
\description{
Read Default Arguments for \code{draft_report()} from YAML-file
}
\examples{
path <- write_default_draft_report_args(path=tempfile(fileext=".yaml"))
read_default_draft_report_args(path=path)
}
