% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_checkbox_sets.R
\name{recode_checkbox_sets}
\alias{recode_checkbox_sets}
\title{Recodes Multiple Sets of Factor Variables of Checkbox Items}
\usage{
recode_checkbox_sets(
  data,
  var_patterns = NULL,
  input_check_value = 1L,
  input_uncheck_value = NA_integer_,
  output_check_value = 2L,
  output_uncheck_value = 1L,
  output_all_unchecked_value = NA_integer_,
  output_check_label = "Selected",
  output_uncheck_label = "Not selected",
  output_all_unchecked_label = NULL,
  convert_to_factor = TRUE,
  keep_tmp_var = FALSE
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{var_patterns}{Character vector of regex search patterns that uniquely identify
variables belonging to sets of batteries. Same as used in tidyselect's matches().}

\item{input_check_value, input_uncheck_value}{Factor/integer values indicating checked and unchecked responses.}

\item{output_check_value, output_uncheck_value}{Recoded values for checked and unchecked responses.}

\item{output_all_unchecked_value}{If respondent has not checked any checkboxes, what should they be recoded as? Set to NULL to not recode.}

\item{output_check_label, output_uncheck_label, output_all_unchecked_label}{Factor labels.}

\item{convert_to_factor}{Flag. Whether to convert variables to factor (using \verb{output_*_label} args) or leave as is.}

\item{keep_tmp_var}{Whether to keep temporary variables for output_all_unchecked_value (mostly for debugging). Defaults to FALSE.}
}
\value{
Data frame/tibble.
}
\description{
Qualtrics produces datasets where a matrix set of checkbox questions
have value labels (factor labels) representing the actual question. When presenting
these questions in a figure or table, however, it is often useful to recode
the value labels to something common, e.g. "Selected" and "Not selected".
}
\details{
Furthermore, this function recodes all of the variables within a set to missing
if all of the variables are missing - under the assumption that the respondent
has then omitted them all. In other cases, e.g. one checkbox is checked, then
the other variables can be assumed to be consciously "Not selected".
}
\examples{
n <- 100
set.seed(1241)
tmp <-
data.frame(a_1 = factor(sample(c("Car", NA), size = n, replace = TRUE)),
           a_2 = factor(sample(c("Train", NA), size = n, replace = TRUE)),
           a_3 = factor(sample(c("Bus", NA), size = n, replace = TRUE)),
           a_4 = factor(sample(c("Flight", NA), size = n, replace = TRUE)),
           b_1 = factor(sample(c("Apple", NA), size = n, replace = TRUE)),
           b_2 = factor(sample(c("Orange", NA), size = n, replace = TRUE)),
           b_3 = factor(sample(c("Banana", NA), size = n, replace = TRUE)),
           b_4 = factor(sample(c("Grapes", NA), size = n, replace = TRUE)))
tmp2 <- recode_checkbox_sets(tmp, var_patterns = c("^a_", "^b_[[:digit:]]$"))
cbind(tmp, tmp2)

}
