% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{refine_data_overview}
\alias{refine_data_overview}
\title{Processes A 'data_overview' Data Frame}
\usage{
refine_data_overview(
  data_overview,
  data = NULL,
  label_separator = NULL,
  name_separator = NULL,
  ...,
  call = rlang::caller_env()
)
}
\arguments{
\item{data_overview}{A data frame containing the columns:}

\item{data}{Optional full survey data set for which columns can be looked up.}

\item{label_separator}{Optional string for separating label between main question and sub-item.}

\item{name_separator}{Optional string for separating column name between main question and sub-item.}

\item{groupby}{Character vector of colnames used for identifying chapters and sections.}

\item{sort_by}{Character vector of of colnames used to order pieces within each groupby combination.}
}
\value{
Data frame
}
\description{
Processes A 'data_overview' Data Frame
}
\examples{
ref_df <- refine_data_overview(ex_survey_ch_overview)
ref_df2 <- refine_data_overview(ex_survey_ch_overview,
                     data = ex_survey1,
                     label_separator = " - ",
                     hide_bi_entry_if_sig_above = .05,
                     name_separator = "_")
}
