% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_label_part.R
\name{remove_label_parts}
\alias{remove_label_parts}
\title{Remove regex pattern from variable labels}
\usage{
remove_label_parts(data, pattern = NULL, data_type_criterion = NULL)
}
\arguments{
\item{data}{Data.frame}

\item{pattern}{String, regex pattern.}

\item{data_type_criterion}{Can be a function that specifies what kind of variables to select, e.g. is.factor}
}
\value{
Data
}
\description{
Remove regex pattern from variable labels
}
\examples{
remove_label_parts(ex_survey1, pattern = " living in")
}
