% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{remove_special_chars_in_labels}
\alias{remove_special_chars_in_labels}
\title{Remove Special Characters (<,>) in Variable Labels}
\usage{
remove_special_chars_in_labels(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
A data frame
}
\description{
Remove Special Characters (<,>) in Variable Labels
}
