% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_by_labels.R, R/utils_labels.R
\name{rename_by_labels}
\alias{rename_by_labels}
\title{Rename Dataset Columns by Labels.}
\usage{
rename_by_labels(
  data,
  label_sep = " - ",
  sort_var = c("pos", "variable", "label"),
  new_var_sep = "_",
  stop_words = NULL
)

rename_by_labels(
  data,
  label_sep = " - ",
  sort_var = c("pos", "variable", "label"),
  new_var_sep = "_",
  stop_words = NULL
)
}
\arguments{
\item{data}{Dataset.}

\item{label_sep}{The separator between group part and unique part of label.}

\item{sort_var}{When numbering variables within a group, what to sort by? pos is original position in dataset,}

\item{new_var_sep}{When creating new variables, how to glue together variable group name prefix and numbering?}

\item{stop_words}{Words to ignore in label when abbreviating label to name.}
}
\value{
Data with renamed variable names.

Data with renamed variable names.
}
\description{
Occasionally dataframe columns have not been named logically and consistent in the
software where the data originates. This function renames variable names
based on patterns in the variable labels, after ignoring some stop words.

Occasionally dataframe columns have not been named logically and consistent in the
software where the data originates. This function renames variable names
based on patterns in the variable labels, after ignoring some stop words.
}
\examples{
rename_by_labels(ex_survey)
rename_by_labels(ex_survey)
}
