% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_saros_report.R
\name{render_saros_report}
\alias{render_saros_report}
\title{Automatically Render a Saros Report from Scratch}
\usage{
render_saros_report(
  chapter_overview,
  data,
  path = "testreport",
  report_generation_yaml_path = fs::path(path, "_report_generation_setup.yml")
)
}
\arguments{
\item{chapter_overview}{A raw overview of chapters, as a data frame with one row per chapter. Must contain columns 'chapter', 'dep_cat', etc.}

\item{data}{A data frame (or a srvyr-object) with the columns specified in the chapter_overview 'dep_cat', etc columns.}

\item{path}{A path, defaults to working directory.}

\item{yaml_path}{Path to a yml-file specifying the}
}
\value{
Path to index qmd-file. If not specified in the yaml_path file, will default to index.qmd.
}
\description{
Automatically Render a Saros Report from Scratch
}
\examples{
\dontrun{
render_saros_report(
    chapter_overview = ex_survey_ch_overview,
    data = ex_survey1)
}
}
