% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_reordered.R
\name{reorder_within}
\alias{reorder_within}
\title{Code-snippets copied and modified from tidytext-package
\url{https://github.com/juliasilge/tidytext/blob/main/R/reorder_within.R}}
\source{
"Original: Ordering categories within ggplot2 Facets" by Tyler Rinker:
\url{https://trinkerrstuff.wordpress.com/2016/12/23/ordering-categories-within-ggplot2-facets/}
Based on \url{https://opensource.org/licenses/MIT}
Copyright (c) 2017, Julia Silge and David Robinson
}
\usage{
reorder_within(x, by, within, fun = mean, sep = "___", ...)
}
\arguments{
\item{x}{Vector}

\item{by}{Vector}

\item{within}{Vector (factor)}

\item{fun}{Function, defaults to the mean}

\item{sep}{String, separator}

\item{...}{Dots}
}
\description{
Permission is hereby granted, free of charge, to any person obtaining
a copy of this software and associated documentation files (the
"Software"), to deal in the Software without restriction, including
without limitation the rights to use, copy, modify, merge, publish,
distribute, sublicense, and/or sell copies of the Software, and to
permit persons to whom the Software is furnished to do so, subject to
the following conditions:
}
\details{
The above copyright notice and this permission notice shall be
included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
}
\keyword{internal}
