% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_replace_docx_imgs_with_mscharts.R
\name{replace_docx_imgs_with_mscharts}
\alias{replace_docx_imgs_with_mscharts}
\title{Replace Images in A Quarto-Generated docx-file with Mschart docx-files}
\usage{
replace_docx_imgs_with_mscharts(file)
}
\arguments{
\item{file}{String path to a docx-file which has images that will be replaced}
}
\value{
String, path to input file.
}
\description{
Replace Images in A Quarto-Generated docx-file with Mschart docx-files
}
