% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{sanitize_labels}
\alias{sanitize_labels}
\title{Sanitize labels originating from e.g. Qualtrics for use in Saros}
\usage{
sanitize_labels(data, sep = " - ", multi_sep_replacement = ": ")
}
\arguments{
\item{data}{data.frame or tibble}

\item{sep}{String, separates main question from subquestion}

\item{multi_sep_replacement}{String. If multiple sep are found, replace the first ones with this.}
}
\value{
Identical data.frame as input, with only variable labels changed.
}
\description{
This function is quite specific to a few problems, users might find it lacking in functionality.
}
\examples{
sanitize_labels(ex_survey1)
}
