% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{saros_conflicts}
\alias{saros_conflicts}
\title{Conflicts between the saros and other packages}
\usage{
saros_conflicts(only = NULL)
}
\arguments{
\item{only}{Set this to a character vector to restrict to conflicts only
with these packages.}
}
\description{
This function lists all the conflicts between packages in saros
and other packages that you have loaded.
}
\examples{
saros_conflicts()
}
