% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saros_packages}
\alias{saros_packages}
\title{List all packages in saros}
\usage{
saros_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include saros in the list?}
}
\description{
List all packages in saros
}
\examples{
saros_packages()
}
