% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest.R
\name{sigtest}
\alias{sigtest}
\title{Test Significance Based on Randomization Theory}
\usage{
sigtest(
  data,
  dep,
  ...,
  indep = NULL,
  .variable_type = NULL,
  indep_type = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{\emph{Survey data}

\verb{obj:<data.frame>|obj:<tbl_df>} // Required

A data frame (or a srvyr-object) with the columns specified in the
chapter_overview 'dep_cat', etc columns.}

\item{dep, indep}{\emph{Variable selections}

<\code{tidyselect}> // \emph{Default:} \code{NULL}, meaning everything for dep, nothing for indep.

Columns in \code{data}. Currently allows tidyselect-syntax, which will be removed.}

\item{...}{\emph{Dynamic dots}

<\href{https://rlang.r-lib.org/reference/dyn-dots.html}{\code{dynamic-dots}}>

Arguments forwarded to the corresponding functions that create the elements.}

\item{.variable_type, indep_type}{Optional string specifying data type ("int", "fct", "character", "factor", "numeric", "integer"). Can be obtained from data lookup.}

\item{call}{\emph{Internal call}

\verb{obj:<call>} // \emph{Default:} \code{rlang::caller_env()} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}
}
\value{
Data frame
}
\description{
Test Significance Based on Randomization Theory
}
\examples{
\dontrun{
set.seed(1)
library(dplyr)
bind_rows(
sigtest(data=ex_survey1, dep = a_1) # t-test of proportions (.5): p=.98
sigtest(data=ex_survey1, dep = b_1), # Chi-square (all=.333): p=<.001
sigtest(data=ex_survey1, dep = c_1), # one-sample t-test p<.001
sigtest(data=ex_survey1, dep = b_1, indep = x1_sex), # Chi-square: p = .548
sigtest(data=ex_survey1, dep = b_1, indep = f_uni), # Chi-square: p = .102
sigtest(data=ex_survey1, dep = c_1, indep = x1_sex), # two-sample t-test: p = .97
sigtest(data=ex_survey1, dep = c_1, indep = f_uni), # ANOVA/F-test: p = .22
sigtest(data=ex_survey1, dep = c_1, indep = c_2)) # correlation: p = .976

sigtest(data=ex_survey1, dep = a_1, indep = c_1) # NA
sigtest(data=ex_survey1, dep = b_1, indep = c_1) # NA
}

}
