% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest_table.R
\name{sigtest_table}
\alias{sigtest_table}
\title{Make Table with All Combinations of Univariate/Bivariate Significance Tests
Based on Variable Types}
\usage{
sigtest_table(data, dep, indep = NULL)
}
\arguments{
\item{data}{Data frame (or tibble)}

\item{dep, indep}{Character vector of dependent/independent variable names.}
}
\value{
Data.frame
}
\description{
Although there are hundreds of significance tests for associations between
two variables, depending upon the distributions, variables types and
assumptions, most fall into a smaller set of popular tests. This function
runs for all combinations of dependent and independent variables in data,
with a suitable test (but not the only possible) for the combination. Also
supports univariate tests, where the assumptions are that of a mean of zero
for continuous variables or all equal proportions for binary/categorical.
}
\details{
This function does not allow any adjustments - use the original underlying
functions for that (chisq.test, t.test, etc.)
}
\examples{
sigtest_table(ex_survey, dep=paste0("b_", 1:3), indep="x1_sex")
}
