% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Stopwords}
\format{
Stopwords as list of character vectors
\describe{contains a list of 58 languages}
}
\usage{
stopwords
}
\description{
Stopwords
}
\keyword{internal}
