% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_vector}
\alias{subset_vector}
\title{Given Ordered Integer Vector, Return Requested Set.}
\usage{
subset_vector(
  vec,
  set = c(".top", ".upper", ".mid_upper", ".lower", ".mid_lower", ".bottom", ".spread"),
  spread_n = NULL,
  sort = FALSE
)
}
\arguments{
\item{vec}{A vector of any type.}

\item{set}{A character string, one of c(".top", ".upper", ".mid_upper", ".lower",
".mid_lower", ".bottom")}

\item{spread_n}{The number of values to extract when set is "spread".}

\item{sort}{Whether to sort the output, defaults to FALSE.}
}
\value{
Selected set of vector.
}
\description{
Useful for identifying which categories are to be collected.
}
\keyword{internal}
