% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_by_labels.R, R/utils_labels.R
\name{swap_label_colnames}
\alias{swap_label_colnames}
\title{Swap Dataset Columns and Labels}
\usage{
swap_label_colnames(data)

swap_label_colnames(data)
}
\arguments{
\item{data}{Data frame or tibble.}
}
\value{
Data.frame/tibble.

Data.frame/tibble.
}
\description{
Columns not containing labels will remain unaffected, and warning given.

Columns not containing labels will remain unaffected, and warning given.
}
\examples{
swap_label_colnames(mtcars)
swap_label_colnames(mtcars)
}
