% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tab_cat_prop}
\alias{tab_cat_prop}
\title{Tabulate categorical variables with proportions}
\usage{
tab_cat_prop(
  data,
  yvars = NULL,
  xvars = NULL,
  var_labels = 2,
  val_labels = FALSE,
  prop = TRUE,
  prefix = "v",
  na.rm.yvar = FALSE,
  na.rm.xvar = FALSE
)
}
\arguments{
\item{data}{data.frame with only variables you want included.}

\item{yvars}{Dependent variable, as string.}

\item{xvars}{Optional independent variable, as string.}

\item{var_labels}{Integer, 0=variable name, 1=variable name - variable label,
2=variable label.}

\item{val_labels}{Logical, whether to use value labels as column names in output.}

\item{prop}{Logical, whether to compute proportions (default), or n_observations.}

\item{prefix}{Prefix to frequency variables in output.}

\item{na.rm.yvar}{Remove missing in dependent variable? Defaults to false.}

\item{na.rm.xvar}{Remove missing in independent variable? Defaults to false.}
}
\value{
data.frame with some fixed columns
}
\description{
Tabulate categorical variables with proportions
}
\examples{
tab_cat_prop(ex_survey1[,paste0("b_", 1:3)])
tab_cat_prop(ex_survey1, xvars = "x1_sex", yvars = paste0("b_", 1:3))
}
