% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tab_cat_prop_to_table}
\alias{tab_cat_prop_to_table}
\title{Turn tibble-result from tab_cat_prop into a prettier table}
\usage{
tab_cat_prop_to_table(data, prefix = "")
}
\arguments{
\item{data}{data.frame from tab_cat_prop}

\item{prefix}{string as prefix for column names}
}
\value{
data.frame
}
\description{
Does what it says.
}
\examples{
tab_cat_prop_to_table(tab_cat_prop(ex_survey1[,paste0("b_", 1:3)]))
}
