% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multi_comb.R
\name{test_multiple_comb}
\alias{test_multiple_comb}
\title{Extension to the infer package for multiple testing}
\usage{
test_multiple_comb(
  data,
  ...,
  formula = NULL,
  response = NULL,
  explanatory = NULL,
  order = NULL,
  alternative = "two-sided",
  mu = 0,
  correct = TRUE,
  p = NULL,
  conf_int = TRUE,
  conf_level = 0.95,
  test = c("t", "chisq", "prop"),
  success = NULL,
  z = FALSE,
  parametric = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{...}{For passing in other arguments to \link[stats:t.test]{t.test()}.}

\item{formula}{A formula with the response variable on the left and the
explanatory on the right.}

\item{response}{The variable name in \code{x} that will serve as the response.
This is alternative to using the \code{formula} argument.}

\item{explanatory}{The variable name in \code{x} that will serve as the
explanatory variable.}

\item{order}{A string vector of specifying the order in which the levels of
the explanatory variable should be ordered for subtraction, where \code{order = c("first", "second")} means \code{("first" - "second")}.}

\item{alternative}{Character string giving the direction of the alternative
hypothesis. Options are \code{"two-sided"} (default), \code{"greater"}, or \code{"less"}.}

\item{mu}{A numeric value giving the hypothesized null mean value for a one
sample test and the hypothesized difference for a two sample test.}

\item{correct}{A logical indicating whether Yates' continuity correction
should be applied where possible. If \code{z = TRUE}, the \code{correct} argument will
be overwritten as \code{FALSE}. Otherwise defaults to \code{correct = TRUE}.}

\item{p}{A numeric vector giving the hypothesized null proportion of
success for each group.}

\item{conf_int}{A logical value for whether to include the confidence
interval or not. \code{TRUE} by default.}

\item{conf_level}{A numeric value between 0 and 1. Default value is 0.95.}

\item{test}{String indicating type of test done for all combinations.}

\item{success}{The level of \code{response} that will be considered a success, as
a string. Only used when testing the null that a single
proportion equals a given value, or that two proportions are equal;
ignored otherwise.}

\item{z}{A logical value for whether to report the statistic as a standard
normal deviate or a Pearson's chi-square statistic. \eqn{z^2}  is distributed
chi-square with 1 degree of freedom, though note that the user will likely
need to turn off Yates' continuity correction by setting \code{correct = FALSE}
to see this connection.}

\item{parametric}{Logical, if false, applies a set of non-parametric tests equivalent to the t-, chisq-, and prop tests.}

\item{call}{Error call function, usually not needed.}
}
\value{
A tibble containing for each response-explanatory row combination
columns summarizing the test statistic.
}
\description{
Extension to the infer package for multiple testing
}
\examples{
library(dplyr)
x <- test_multiple_comb(data = mtcars,
                     response = matches("mpg|disp|drat|hp|qsec"),
                     explanatory = matches("vs|am"),
                     test = "t")
}
