% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_bool}
\alias{validate_bool}
\title{Validate boolean parameter}
\usage{
validate_bool(x, call = rlang::caller_env(), arg = rlang::caller_arg(x))
}
\description{
Validate boolean parameter
}
\keyword{internal}
