% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_double}
\alias{validate_double}
\title{Validate double/numeric parameter}
\usage{
validate_double(
  x,
  min = -Inf,
  max = Inf,
  null_allowed = FALSE,
  call = rlang::caller_env(),
  arg = rlang::caller_arg(x)
)
}
\arguments{
\item{x}{Value to validate}

\item{min}{Minimum allowed value (default -Inf)}

\item{max}{Maximum allowed value (default Inf)}

\item{null_allowed}{Whether NULL is an acceptable value (default FALSE)}

\item{call}{Calling environment for error messages}

\item{arg}{Argument name for error messages}
}
\description{
Validate double/numeric parameter
}
\keyword{internal}
