% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_double_rule}
\alias{validate_double_rule}
\title{Create double validation rule}
\usage{
validate_double_rule(min = -Inf, max = Inf, null_allowed = FALSE)
}
\arguments{
\item{min}{Minimum allowed value (default -Inf)}

\item{max}{Maximum allowed value (default Inf)}

\item{null_allowed}{Whether NULL is an acceptable value (default FALSE)}
}
\description{
Create double validation rule
}
\keyword{internal}
