% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_draft_report_arguments_in_yaml.R
\name{write_default_draft_report_args}
\alias{write_default_draft_report_args}
\title{Write Default Arguments for \code{draft_report()} to YAML-file}
\usage{
write_default_draft_report_args(path)
}
\arguments{
\item{path}{\verb{scalar<character>} // Required. \emph{default:} \code{settings.yaml}}
}
\value{
The defaults as a \code{yaml}-object.
}
\description{
Write Default Arguments for \code{draft_report()} to YAML-file
}
\examples{
write_default_draft_report_args(path=tempfile(fileext=".yaml"))
}
