% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{add_dep_order}
\alias{add_dep_order}
\title{Add dependent variable ordering}
\usage{
add_dep_order(data, sort_by, descend = FALSE)
}
\arguments{
\item{data}{Dataset}

\item{sort_by}{Sorting method for dependent variables}

\item{descend}{Whether to reverse the order}
}
\value{
Dataset with .dep_order column added
}
\description{
Add dependent variable ordering
}
\keyword{internal}
