% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{apply_final_arrangement}
\alias{apply_final_arrangement}
\title{Apply final arrangement based on order columns}
\usage{
apply_final_arrangement(data)
}
\arguments{
\item{data}{Dataset with order columns}
}
\value{
Arranged dataset
}
\description{
Apply final arrangement based on order columns
}
\keyword{internal}
