% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cat_cat_data.R
\name{apply_legacy_sorting_adjustments}
\alias{apply_legacy_sorting_adjustments}
\title{Apply legacy sorting adjustments for special cases}
\usage{
apply_legacy_sorting_adjustments(
  data,
  indep_names = character(0),
  translations = list()
)
}
\arguments{
\item{data}{Dataset}

\item{indep_names}{Independent variable names}

\item{translations}{Translation strings}
}
\value{
Dataset with legacy adjustments applied
}
\description{
Apply legacy sorting adjustments for special cases
}
\keyword{internal}
