% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{arrange_table_data}
\alias{arrange_table_data}
\title{Arrange output data by prespecified orders}
\usage{
arrange_table_data(data, col_basis, indep_vars = NULL)
}
\arguments{
\item{data}{Data frame to arrange}

\item{col_basis}{Column to use as primary sort}

\item{indep_vars}{Independent variable columns}
}
\value{
Arranged data frame
}
\description{
Standard data arrangement for table functions
}
\keyword{internal}
