% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attach_dep_label_prefix}
\alias{attach_dep_label_prefix}
\title{Attach dep_label_prefix attribute to a make_content output object}
\usage{
attach_dep_label_prefix(obj, main_question)
}
\arguments{
\item{obj}{The object to annotate (ggplot, data.frame, mschart, etc.)}

\item{main_question}{Character scalar; the dep label prefix.}
}
\value{
\code{obj} with \code{attr(obj, "dep_label_prefix")} set (invisibly if NULL/NA).
}
\description{
Attaches the main question (i.e. the label prefix of the dep variables)
as an attribute \code{"dep_label_prefix"} on the returned object. Works for
ggplot, data.frame, and mschart objects. Should not be called when returning
an rdocx object.
}
\keyword{internal}
