% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe-utils.R
\name{build_custom_palette}
\alias{build_custom_palette}
\title{Build Custom Palette Function}
\usage{
build_custom_palette(palette_codes, fct_levels, priority_palette_codes = NULL)
}
\description{
Build Custom Palette Function
}
\keyword{internal}
