% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe-utils.R
\name{build_custom_palette}
\alias{build_custom_palette}
\title{Build Custom Palette Function}
\usage{
build_custom_palette(palette_codes, fct_levels, priority_palette_codes = NULL)
}
\arguments{
\item{palette_codes}{List of character vectors containing colours. Vectors can
optionally be named for exact level matching.}

\item{fct_levels}{Character vector of factor levels to assign colors to.}

\item{priority_palette_codes}{Optional named character vector where names are
categories and values are colours to use first. Defaults to \code{NULL}.}
}
\value{
A palette function that takes n (number of colors) and lvls (levels)
and returns a named character vector of colors.
}
\description{
Creates a palette function that matches colors to factor levels based on
palette_codes and priority_palette_codes.
}
\keyword{internal}
