% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{calculate_indep_column_order}
\alias{calculate_indep_column_order}
\title{Calculate independent variable ordering based on a specific column value}
\usage{
calculate_indep_column_order(
  data,
  column_name,
  indep_col,
  descend_indep = FALSE
)
}
\arguments{
\item{data}{Dataset}

\item{column_name}{Name of the column to sort by}

\item{indep_col}{Name of the independent variable column}

\item{descend_indep}{Logical indicating if sorting should be descending}
}
\value{
Numeric vector of ordering values
}
\description{
Calculate independent variable ordering based on a specific column value
}
\keyword{internal}
