% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{calculate_indep_proportion_order}
\alias{calculate_indep_proportion_order}
\title{Calculate independent variable ordering based on position categories}
\usage{
calculate_indep_proportion_order(
  data,
  method,
  indep_col,
  descend_indep = FALSE
)
}
\arguments{
\item{data}{Dataset}

\item{method}{Either ".upper", ".top", etc.}

\item{indep_col}{Name of the independent variable column}

\item{descend_indep}{Logical indicating if sorting should be descending}
}
\value{
Numeric vector of ordering values
}
\description{
Calculate independent variable ordering based on position categories
}
\keyword{internal}
