% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_utils.R
\name{calculate_indep_sum_value_order}
\alias{calculate_indep_sum_value_order}
\title{Calculate independent variable ordering based on multiple category values}
\usage{
calculate_indep_sum_value_order(
  data,
  category_values,
  indep_col,
  descend_indep = FALSE
)
}
\arguments{
\item{data}{Dataset}

\item{category_values}{Vector of category values to sum}

\item{indep_col}{Name of the independent variable column}

\item{descend_indep}{Logical indicating if sorting should be descending}
}
\value{
Numeric vector of ordering values
}
\description{
Calculate independent variable ordering based on multiple category values
}
\keyword{internal}
