% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest.R
\name{embed_bi_sigtest}
\alias{embed_bi_sigtest}
\title{Bivariate significance tests}
\usage{
embed_bi_sigtest(
  data,
  dep,
  indep,
  ...,
  .variable_type = NULL,
  indep_type = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{\emph{Survey data}

\verb{obj:<data.frame>|obj:<tbl_df>} // Required

A data frame (or a srvyr-object) with the columns specified in the
chapter_overview 'dep_cat', etc columns.}

\item{dep, indep}{\emph{Variable selections}

<\code{tidyselect}> // \emph{Default:} \code{NULL}, meaning everything for dep, nothing for indep.

Columns in \code{data}. Currently allows tidyselect-syntax, which will be removed.}

\item{...}{\emph{Dynamic dots}

<\href{https://rlang.r-lib.org/reference/dyn-dots.html}{\code{dynamic-dots}}>

Arguments forwarded to the corresponding functions that create the elements.}

\item{.variable_type, indep_type}{Optional string specifying data type ("int", "fct", "character", "factor", "numeric", "integer"). Can be obtained from data lookup.}

\item{call}{\emph{Internal call}

\verb{obj:<call>} // \emph{Default:} \code{rlang::caller_env()} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}
}
\value{
Data frame
}
\description{
Bivariate significance tests
}
\examples{
embed_bi_sigtest(data=ex_survey, dep = a_1:a_9, indep = x1_sex,
                 reps=10, hide_test_if_n_below=1)
embed_bi_sigtest(data=ex_survey, dep = b_1:b_3, indep = x1_sex,
                 reps=10, hide_test_if_n_below=1, label_separator=" - ")
embed_bi_sigtest(data=ex_survey, dep = c_1:c_2, indep = x1_sex,
                 reps=10, hide_test_if_n_below=1)
}
