% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_freq_plot_html.R
\name{embed_cat_freq_plot}
\alias{embed_cat_freq_plot}
\title{Embed Interactive Categorical Plot}
\usage{
embed_cat_freq_plot(
  data,
  ...,
  cols = tidyselect::everything(),
  by = NULL,
  summarized_data = NULL,
  label_separator = NULL,
  translations = getOption("saros")$translations,
  html_interactive = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Optional parameters forwarded from above.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{translations}{List of named strings for by and (N=, etc. see getOption("saros")$translations.}

\item{call}{Error call function, usually not needed.}

\item{plot_height_multiplier}{[\code{numeric(1)>0}]\cr Height in cm per chart entry.}

\item{plot_height_fixed_constant}{[\code{numeric(1)>0}]\cr Fixed height in cm.}

\item{return_raw}{[\code{logical(1)}] Whether to return the raw static chart. Defaults to FALSE.}
}
\value{
ggplot
}
\description{
Embed Interactive Categorical Plot
}
\examples{
\dontrun{
embed_cat_freq_plot(data = ex_survey1, cols = b_1:b_3)
}
}
