% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_freq_plot_docx.R
\name{embed_cat_freq_plot_docx}
\alias{embed_cat_freq_plot_docx}
\title{Create Word Report with Univariates for Categorical Columns Sharing Same Categories}
\usage{
embed_cat_freq_plot_docx(
  data,
  ...,
  cols = tidyselect::everything(),
  by = NULL,
  summarized_data = NULL,
  label_separator = NULL,
  tailored_group = NULL,
  translations = .saros.env$defaults$translations
)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Optional parameters forwarded from above.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{summarized_data}{Currently not in use}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{tailored_group}{String, indicating name of tailored group.}

\item{translations}{List of named strings for by and (N=, etc. see getOption("saros")$translations.}
}
\value{
rdocx object, which can be saved with print() after loading the officer-package
}
\description{
Create Word Report with Univariates for Categorical Columns Sharing Same Categories
}
\examples{
library(officer) # To save the rdocx object to disk

 test_docx_b13 <-
   ex_survey1 |>
   embed_cat_freq_plot_docx(cols = b_1:b_3,
              showNA = "never",
              descend = TRUE,
              return_raw = FALSE,
              hide_label_if_prop_below=0,
              data_label = "count",
              data_label_decimal_symbol = ",",
              digits = 1,
              label_font_size = 12,
              main_font_size = 12,
             plot_height_multiplier = .3,
             plot_height_fixed_constant = 1,
              vertical = FALSE,
              font_family = "sans")
\dontrun{
print(test_docx_b13, target = "test_docx_b13.docx")
file.remove("test_docx_b13.docx")
}
}
