% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table.R
\name{embed_cat_table}
\alias{embed_cat_table}
\title{Embed Reactable Table}
\usage{
embed_cat_table(
  data,
  ...,
  dep = tidyselect::everything(),
  indep = NULL,
  mesos_group = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{\emph{Survey data}

\verb{obj:<data.frame>|obj:<tbl_df>} // Required

A data frame (or a srvyr-object) with the columns specified in the
chapter_overview 'dep_cat', etc columns.}

\item{...}{\emph{Dynamic dots}

<\href{https://rlang.r-lib.org/reference/dyn-dots.html}{\code{dynamic-dots}}>

Arguments forwarded to the corresponding functions that create the elements.}

\item{dep, indep}{\emph{Variable selections}

<\code{tidyselect}> // \emph{Default:} \code{NULL}, meaning everything for dep, nothing for indep.

Columns in \code{data}. Currently allows tidyselect-syntax, which will be removed.}

\item{mesos_group}{\emph{Specific group to compare with}

\verb{scalar<character>} // \emph{Default:} \code{NULL} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}

\item{call}{\emph{Internal call}

\verb{obj:<call>} // \emph{Default:} \code{rlang::caller_env()} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}
}
\value{
A reactable object. If return_raw=FALSE, then just the table.
}
\description{
Embed Reactable Table
}
