% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table.R
\name{embed_cat_table}
\alias{embed_cat_table}
\title{Embed Reactable Table (DEPRECATED!)}
\usage{
embed_cat_table(
  data,
  ...,
  dep = tidyselect::everything(),
  indep = NULL,
  mesos_group = NULL
)
}
\arguments{
\item{data}{\code{data.frame}, \code{tibble} or potentially a \code{srvyr}-object.}

\item{...}{Dynamic dots, arguments forwarded to underlying function(s).}

\item{dep}{\code{tidyselect}-syntax for dependent variable(s).}

\item{indep}{\code{tidyselect}-syntax for an optional independent variable.}

\item{mesos_group}{String}
}
\description{
This function has been deprecated.
Use instead \code{\link[saros.contents:makeme]{saros.contents::makeme()}}
}
