% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{embed_cat_text_html}
\alias{embed_cat_text_html}
\title{Creates a structured list with text interpretations for a set of variables.}
\usage{
embed_cat_text_html(
  data,
  ...,
  cols = NULL,
  by = NULL,
  summarized_data = NULL,
  label_separator = NULL,
  translations = .saros.env$defaults$translations,
  tailored_group = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Optional parameters forwarded from above.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{translations}{A list of translations for the template text. See .saros.env$defaults.}

\item{tailored_group}{String, indicating the name of the tailored group}

\item{call}{Error call function, usually not needed.}

\item{contents}{The type of text interpretations to return, multiple allowed. Defaults to all.}

\item{include_numbers}{Whether or not to include the actual numbers (percentages, means) in parentheses.}

\item{require_common_categories}{Whether to check if all questions share common categories.}

\item{n_top_bottom}{The number of top and bottom entries to report.}
}
\value{
List
}
\description{
Creates a structured list with text interpretations for a set of variables.
}
\examples{
embed_cat_text_html(ex_survey1,
cols = tidyselect::matches("e_"),
contents = c("intro", "mode_max", "value_max",
"value_min", "not_used_category", "mean_max", "mean_min"),
label_separator = " - ",
require_common_categories = FALSE,
n_top_bottom = 1,
showNA = "never",
descend = TRUE,
return_raw = TRUE,
hide_label_if_prop_below=0,
data_label = "count",
data_label_decimal_symbol = ",",
digits = 1)

}
