% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest.R
\name{embed_sigtest}
\alias{embed_sigtest}
\title{Title}
\usage{
embed_sigtest(
  data,
  cols,
  ...,
  by = NULL,
  col_type = NULL,
  by_type = NULL,
  reps = 1000,
  digits = 1,
  return_raw = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{...}{Optional parameters forwarded from above.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{col_type, by_type}{Optional string specifying data type ("int", "cat", "character", "factor", "numeric", "integer"). Can be obtained from data lookup.}

\item{reps}{Integer, number of permutations.}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{return_raw}{[\code{logical(1)}] Whether to return the raw static chart. Defaults to FALSE.}

\item{call}{Error call function, usually not needed.}
}
\value{
Data frame
}
\description{
Title
}
