% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest.R
\name{embed_uni_sigtest}
\alias{embed_uni_sigtest}
\title{Run univariate significance tests and collect them in a table}
\usage{
embed_uni_sigtest(
  data,
  dep,
  ...,
  indep = NULL,
  .variable_type = NULL,
  indep_type = NULL,
  mesos_group = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{\emph{Survey data}

\verb{obj:<data.frame>|obj:<tbl_df>} // Required

A data frame (or a srvyr-object) with the columns specified in the
chapter_overview 'dep_cat', etc columns.}

\item{dep, indep}{\emph{Variable selections}

<\code{tidyselect}> // \emph{Default:} \code{NULL}, meaning everything for dep, nothing for indep.

Columns in \code{data}. Currently allows tidyselect-syntax, which will be removed.}

\item{...}{\emph{Dynamic dots}

<\href{https://rlang.r-lib.org/reference/dyn-dots.html}{\code{dynamic-dots}}>

Arguments forwarded to the corresponding functions that create the elements.}

\item{.variable_type, indep_type}{Optional string specifying data type ("int", "fct", "character", "factor", "numeric", "integer"). Can be obtained from data lookup.}

\item{mesos_group}{\emph{Specific group to compare with}

\verb{scalar<character>} // \emph{Default:} \code{NULL} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}

\item{call}{\emph{Internal call}

\verb{obj:<call>} // \emph{Default:} \code{rlang::caller_env()} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}
}
\value{
Data frame
}
\description{
Run univariate significance tests and collect them in a table
}
\examples{
embed_uni_sigtest(data=ex_survey,
                  dep = a_1:a_9, reps=10, hide_test_if_n_below=10)
embed_uni_sigtest(data=ex_survey,
                  dep = b_1:b_3, reps=10, hide_test_if_n_below=10,
                  label_separator=" - ")
embed_uni_sigtest(data=ex_survey,
                  dep = c_1:c_2, reps=10, hide_test_if_n_below=10)
}
