% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_survey_ch_overview}
\alias{ex_survey_ch_overview}
\title{ex_survey_ch_overview: Mock overview of chapter structure}
\format{
A data frame with 5 rows (chapters) and 5 variables:
\describe{
\item{chapter}{Manual entry chapter title}
\item{author}{Single or multiple authors, separated by comma}
\item{dep}{Columns in ex_survey having the role of dependent variable}
\item{indep}{Columns in ex_survey having the role of independent variable}
\item{irrelevant_col}{Just a column about something else to verify that the system works also with superfluous information.}
}
}
\usage{
ex_survey_ch_overview
}
\description{
Note that only chapter and dep are compulsory.
}
\keyword{datasets}
