% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_survey_ch_overview}
\alias{ex_survey_ch_overview}
\title{ex_survey_ch_overview: Mock overview of chapter structure}
\format{
A data frame with 3 rows and 6 variables:
\describe{
\item{chapter}{Manual entry chapter title}
\item{author}{Single or multiple authors, separated by comma}
\item{dep}{Columns in ex_survey1 having the role of dependent variable}
\item{indep}{Columns in ex_survey1 having the role of independent variable}
\item{irrelevant_col}{Just a column about something else to verify that the system works also with superfluous information.}
}
}
\usage{
ex_survey_ch_overview
}
\description{
Note that indep_int, dep_txt, indep_txt and other columns are also welcome.
}
\keyword{datasets}
