% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_height_h_barchart.R
\name{fig_height_h_barchart2}
\alias{fig_height_h_barchart2}
\title{Estimate figure height for a horizontal bar chart}
\usage{
fig_height_h_barchart2(
  ggobj,
  freq = FALSE,
  x_axis_label_width = 20,
  strip_angle = 0,
  main_font_size = 8,
  legend_location = c("panel", "plot"),
  n_legend_lines = 2,
  legend_key_chars_equivalence = 5,
  max_chars_per_figure_width = 100,
  multiplier_per_horizontal_line = NULL,
  multiplier_per_vertical_letter = 0.01,
  multiplier_per_facet = 0.95,
  multiplier_per_legend_line = 1.5,
  fixed_constant = 0,
  figure_width_in_cm = 16,
  margin_in_cm = 0,
  max = 8,
  min = 1
)
}
\arguments{
\item{ggobj}{\code{ggplot2}-object}

\item{freq}{Logical. If TRUE, frequency plot with categories next to each other. If FALSE (default), proportion plot with stacked categories.}

\item{x_axis_label_width}{Numeric. Width allocated for x-axis labels.}

\item{strip_angle}{Integer. Angle of the strip text.}

\item{main_font_size}{Numeric. Font size for the main text.}

\item{legend_location}{Character. Location of the legend ("panel" or "plot").}

\item{n_legend_lines}{Integer. Number of lines in the legend.}

\item{legend_key_chars_equivalence}{Integer. Approximate number of characters the legend key equals.}

\item{max_chars_per_figure_width}{Integer. Maximum number of characters per figure width.}

\item{multiplier_per_horizontal_line}{Numeric. Multiplier per horizontal line.}

\item{multiplier_per_vertical_letter}{Numeric. Multiplier per vertical letter.}

\item{multiplier_per_facet}{Numeric. Multiplier per facet.}

\item{multiplier_per_legend_line}{Numeric. Multiplier per legend line.}

\item{fixed_constant}{Numeric. Fixed constant to be added to the height.}

\item{figure_width_in_cm}{Numeric. Width of the figure in centimeters.}

\item{margin_in_cm}{Numeric. Margin in centimeters.}

\item{max}{Numeric. Maximum height.}

\item{min}{Numeric. Minimum height.}
}
\value{
Numeric value representing the estimated height of the figure.
}
\description{
Taking an object from \code{makeme()}, this function estimates the height of a
figure for a horizontal bar chart.
}
\examples{
fig_height_h_barchart2(makeme(data=ex_survey, dep=b_1:b_3, indep=x1_sex))
}
