% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_sanitizer.R
\name{filename_sanitizer}
\alias{filename_sanitizer}
\title{File/folder name sanitizer replacing space and punctuation with underscore}
\usage{
filename_sanitizer(x, max_chars = NA_integer_, accept_hyphen = FALSE)
}
\arguments{
\item{x}{Character vector of file/folder names}

\item{max_chars}{Maximum character length}

\item{accept_hyphen}{Flag, whether a hyphen - is acceptable.}
}
\value{
Character vector of same length as x
}
\description{
File/folder name sanitizer replacing space and punctuation with underscore
}
\examples{
filename_sanitizer(c("Too long a name", "with invalid *^/&#"))
}
