% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{get_colour_set}
\alias{get_colour_set}
\title{Provide A Colour Set for A Number of Requested Colours}
\usage{
get_colour_set(
  n_colours_needed,
  user_colour_set = NULL,
  names = NULL,
  colour_na = NULL,
  colour_2nd_binary_cat = NULL,
  seed = 1,
  call = rlang::caller_env()
)
}
\arguments{
\item{n_colours_needed}{Number of colours needed.}

\item{user_colour_set}{User-supplied default palette.}

\item{names}{Optional names of colour vector, follows same options as rlang::set_names.}

\item{colour_na}{[\code{character(1)}]\cr Colour as a single string.}

\item{colour_2nd_binary_cat}{[\code{character(1)}]\cr Colour for second category in binary variables. Often useful to hide this.}

\item{seed}{Random seed for sampling.}

\item{call}{Caller function if used from inside another function.}
}
\value{
A colour set as character vector.
}
\description{
Possibly using user_colour_set if available. If not sufficient, uses a set
    palette from RColorBrewer.
}
\examples{
get_colour_set(n_colours_needed=4)
}
