% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dep_display_column}
\alias{get_dep_display_column}
\title{Determine display column for dependent variables in int_plot_html}
\usage{
get_dep_display_column(dep_count, dep_labels)
}
\arguments{
\item{dep_count}{Number of dependent variables}

\item{dep_labels}{Vector of dependency labels}
}
\value{
Character string indicating which column to use
}
\description{
Checks if the number of dep variables matches the number of labels to determine
whether to use .variable_label or .variable_name for display.
}
\keyword{internal}
