% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handpick.R
\name{handpick}
\alias{handpick}
\title{Return character vector of manually picked data columns.}
\usage{
handpick(data, to_clipboard = FALSE)
}
\arguments{
\item{data}{A dataset; tibble or data frame.}

\item{to_clipboard}{Boolean. Defaults to FALSE. If TRUE, overwrites what you already have copied. Use with caution.}
}
\value{
Character vector. Prints to console.
}
\description{
Only works if in interactive session. Copies to the clipboard if to_clipboard = TRUE, and if clipboard is available on system.
}
\examples{
\dontrun{if(interactive()) handpick(mtcars)}
}
