% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{hex_bw}
\alias{hex_bw}
\title{Identify Suitable Font Given Background Hex Colour}
\usage{
hex_bw(hex_code, na_colour = "#ffffff")
}
\arguments{
\item{hex_code}{Colour in hex-format.}
}
\value{
Colours in hex-format, either black or white.
}
\description{
Code is taken from XXX.
}
\keyword{internal}
