% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{hex_bw}
\alias{hex_bw}
\title{Identify Suitable Font Given Background Hex Colour}
\usage{
hex_bw(hex_code, colour_2nd_binary_cat = NULL)
}
\arguments{
\item{hex_code}{Colour in hex-format.}

\item{colour_2nd_binary_cat}{If hex_code equals this, return hex_code as is.}
}
\value{
Colours in hex-format, either black or white.
}
\description{
Code is taken from XXX.
}
\examples{
hex_bw("#0dadfd")
}
