% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_int_plot_html}
\alias{is_int_plot_html}
\title{Detect if data is from int_plot_html}
\usage{
is_int_plot_html(data)
}
\arguments{
\item{data}{Data frame to check}
}
\value{
Logical indicating if this is int_plot_html format
}
\description{
Checks if a data frame has the structure expected from int_plot_html output,
which uses .value column instead of .category column.
}
\keyword{internal}
