% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_rng.R
\name{n_range2}
\alias{n_range2}
\title{Provides a range (or single value) for N in a \code{ggplot2}-object from \code{makeme()}}
\usage{
n_range2(ggobj, glue_template_1 = "{n}", glue_template_2 = "[{n[1]}-{n[2]}]")
}
\arguments{
\item{ggobj}{A \code{ggplot2}-object.}

\item{glue_template_1, glue_template_2}{String, for the case of a single
value (1) or a range with minimum-maximum of values (2).}
}
\value{
String.
}
\description{
Provides a range (or single value) for N in a \code{ggplot2}-object from \code{makeme()}
}
\examples{
n_range2(makeme(data = ex_survey, dep = b_1:b_3))
}
