% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeme_helpers.R
\name{normalize_makeme_arguments}
\alias{normalize_makeme_arguments}
\title{Normalize Multi-Choice Arguments to Single Values}
\usage{
normalize_makeme_arguments(args)
}
\arguments{
\item{args}{List of makeme function arguments}
}
\value{
Modified args list with normalized single-value arguments:
\itemize{
\item \code{showNA}: First element of showNA vector
\item \code{data_label}: First element of data_label vector
\item \code{data_label_position}: First element of data_label_position vector
\item \code{type}: First element of evaluated type expression
}
}
\description{
Internal helper function that ensures makeme arguments that might be vectors
are normalized to single values by taking the first element.
}
\keyword{internal}
