% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_freq_plot_docx.R
\name{prep_cat_freq_plot_docx}
\alias{prep_cat_freq_plot_docx}
\title{Create Categorical Data Chart from Summarized Data}
\usage{
prep_cat_freq_plot_docx(data, ..., call = rlang::caller_env())
}
\arguments{
\item{data}{Data frame.}

\item{...}{Optional parameters forwarded from above.}

\item{call}{For internal use}
}
\value{
mschart-object. Can be added to an rdocx, rpptx or rxlsx object.
}
\description{
Create Categorical Data Chart from Summarized Data
}
