% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_prop_plot_html.R
\name{prep_cat_prop_plot_html}
\alias{prep_cat_prop_plot_html}
\title{Create Single Interactive Categorical Plot with Univariates for Categorical Columns Sharing Same Categories}
\usage{
prep_cat_prop_plot_html(data, ..., inverse = FALSE, call = rlang::caller_env())
}
\arguments{
\item{data}{\emph{Survey data}

\verb{obj:<data.frame>|obj:<tbl_df>} // Required

A data frame (or a srvyr-object) with the columns specified in the
chapter_overview 'dep_cat', etc columns.}

\item{...}{\emph{Dynamic dots}

<\href{https://rlang.r-lib.org/reference/dyn-dots.html}{\code{dynamic-dots}}>

Arguments forwarded to the corresponding functions that create the elements.}

\item{inverse}{Flag, defaults to FALSE. If TRUE, swaps x-axis and faceting.}

\item{call}{\emph{Internal call}

\verb{obj:<call>} // \emph{Default:} \code{rlang::caller_env()} (\code{optional})

Both the absolute and relative folderpaths are required, as strings.}
}
\value{
ggiraph object, plottable with plot()
}
\description{
Create Single Interactive Categorical Plot with Univariates for Categorical Columns Sharing Same Categories
}
\keyword{internal}
