% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{refine_data_overview}
\alias{refine_data_overview}
\title{Processes A 'data_overview' Data Frame
Take Kapitteloversikt.xlsx, read it in, and send it to here.}
\usage{
refine_data_overview(
  data_overview,
  data = NULL,
  group_by = c("chapter", "col_group", "label_prefix"),
  sort_by = NULL,
  label_separator = NULL,
  name_separator = NULL
)
}
\arguments{
\item{data_overview}{A data frame containing the columns:}

\item{data}{Optional full survey data set for which columns can be looked up.}

\item{group_by}{Character vector of colnames used for identifying chapters and sections.}

\item{sort_by}{Character vector of of colnames used to order pieces within each group_by combination.}

\item{label_separator}{Optional string for separating label between main question and sub-item.}

\item{name_separator}{Optional string for separating column name between main question and sub-item.}
}
\value{
Data frame
}
\description{
Processes A 'data_overview' Data Frame
Take Kapitteloversikt.xlsx, read it in, and send it to here.
}
\examples{
saros::refine_data_overview(ex_survey_ch_overview)
}
