% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_replace_docx_imgs_with_mscharts.R
\name{replace_docx_imgs_with_mscharts}
\alias{replace_docx_imgs_with_mscharts}
\title{Replace Images in A Quarto-Generated docx-file with Mschart docx-files}
\usage{
replace_docx_imgs_with_mscharts(
  main_file,
  chart_dir = fs::path_ext_remove(main_file),
  delete_mschart_files = FALSE
)
}
\arguments{
\item{main_file}{docx file where images are to be replaced with mscharts}

\item{chart_dir}{Directory path to where the mscharts as docx files can be found.}

\item{delete_mschart_files}{Flag. Whether to delete the mschart docx files after insertion. Defaults to FALSE.}
}
\value{
String, path to input file.
}
\description{
Replace Images in A Quarto-Generated docx-file with Mschart docx-files
}
