% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{sanitize_labels}
\alias{sanitize_labels}
\title{Sanitize labels originating from e.g. Qualtrics for use in Saros}
\usage{
sanitize_labels(
  data,
  sep = " - ",
  multi_sep_replacement = ": ",
  replace_ascii_with_utf = FALSE,
  questions = NULL
)
}
\arguments{
\item{data}{data.frame or tibble}

\item{sep}{String, separates main question from subquestion}

\item{multi_sep_replacement}{String. If multiple sep are found, replace the first ones with this.}

\item{replace_ascii_with_utf}{Flag. If TRUE, downloads a list from W3 used to convert html characters as ASCII to UTF8.}

\item{questions}{Data frame with questions obtained from \code{qualtRics::survey_questions()}}
}
\value{
Identical data.frame as input, with only variable labels changed.
}
\description{
This function is quite specific to a few problems, users might find it lacking in functionality.
}
\examples{
sanitize_labels(ex_survey)
}
