% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saros.R
\docType{package}
\name{saros-package}
\alias{saros}
\alias{saros-package}
\title{saros: Semi-Automatic Reporting of Ordinary Surveys}
\description{
Offers a systematic way for conditional reporting of figures and tables for many (and bivariate combinations of) variables, typically from survey data. Contains interactive 'ggiraph'-based (\url{https://CRAN.R-project.org/package=ggiraph}) plotting functions and data frame-based summary tables (bivariate significance tests, frequencies/proportions, unique open ended responses, etc) with many arguments for customization, and extensions possible. Uses a global options() system for neatly reducing redundant code. Also contains tools for immediate saving of objects and returning a hashed link to the object, useful for creating download links to high resolution images upon rendering in 'Quarto'. Suitable for highly customized reports, primarily intended for survey research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nifu-no.github.io/saros/}
  \item \url{https://github.com/NIFU-NO/saros}
  \item Report bugs at \url{https://github.com/NIFU-NO/saros/issues}
}

}
\author{
\strong{Maintainer}: Stephan Daus \email{stephus.daus@gmail.com} (\href{https://orcid.org/0000-0003-0230-6997}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Julia Silge (Author of internal scale_x_reordered) [contributor]
  \item David Robinson (Author of internal scale_x_reordered) [contributor]
  \item Nordic Institute for The Studies of Innovation, Research and Education (NIFU) [funder]
  \item Kristiania University College [funder]
}

}
\keyword{internal}
