% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saros.R
\docType{package}
\name{saros-package}
\alias{saros}
\alias{saros-package}
\title{saros: Semi-Automatic Reporting of Ordinary Surveys}
\description{
Produces highly customized reports, primarily intended for survey research. Building on 'Quarto' (\url{https://quarto.org}), it generates draft chapters of all specified dependent/independent variables, which can be further edited by hand, containing figures, tables and analyses (currently only uni-/bivariate tests of equivalent means/proportions). The feature 'mesos'-reports offer tailor-made batch report production where e.g. an institution can compare itself to all other participants. Publication tools are also included, such as password generation for 'mesos'-report sections on a 'Quarto' Website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nifu-no.github.io/saros/}
  \item \url{https://github.com/NIFU-NO/saros}
  \item Report bugs at \url{https://github.com/NIFU-NO/saros/issues}
}

}
\author{
\strong{Maintainer}: Stephan Daus \email{stephus.daus@gmail.com} (\href{https://orcid.org/0000-0003-0230-6997}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Nordic Institute for The Studies of Innovation, Research and Education (NIFU) [funder]
  \item Kristiania University College [funder]
}

}
\keyword{internal}
