% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{setup_table_data}
\alias{setup_table_data}
\title{Setup table data from dots}
\usage{
setup_table_data(dots)
}
\arguments{
\item{dots}{List from rlang::list2(...)}
}
\value{
List with data and should_return flag
}
\description{
Common setup logic for table functions including data extraction and early return
}
