% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cat_cat_data.R
\name{shift_labels_always_at}
\alias{shift_labels_always_at}
\title{Shift labels_always_at}
\usage{
shift_labels_always_at(data, labels_always_at = NULL, after = Inf)
}
\arguments{
\item{data}{Dataset}

\item{labels_always_at}{Labels to move to bottom or top}

\item{after}{Position to move labels to (0 = top, Inf = bottom)}
}
\value{
Dataset with data$.variable_label adjusted
}
\description{
Shift labels_always_at
}
\keyword{internal}
