% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest.R
\name{sigtest}
\alias{sigtest}
\title{Test Significance Based on Randomization Theory}
\usage{
sigtest(
  data,
  cols,
  ...,
  by = NULL,
  col_type = NULL,
  by_type = NULL,
  label_separator = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{...}{Optional parameters forwarded from above.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{col_type, by_type}{Optional string specifying data type ("int", "fct", "character", "factor", "numeric", "integer"). Can be obtained from data lookup.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{call}{Error call function, usually not needed.}

\item{reps}{Integer, number of permutations.}

\item{hide_test_if_n_below}{Integer, if N is below this value, pvalue will not be shown.}
}
\value{
Data frame
}
\description{
Test Significance Based on Randomization Theory
}
\examples{
\dontrun{
set.seed(1)
library(dplyr)
bind_rows(
sigtest(data=ex_survey1, cols = a_1) # t-test of proportions (.5): p=.98
sigtest(data=ex_survey1, cols = b_1), # Chi-square (all=.333): p=<.001
sigtest(data=ex_survey1, cols = c_1), # one-sample t-test p<.001
sigtest(data=ex_survey1, cols = b_1, by = x1_sex), # Chi-square: p = .548
sigtest(data=ex_survey1, cols = b_1, by = f_uni), # Chi-square: p = .102
sigtest(data=ex_survey1, cols = c_1, by = x1_sex), # two-sample t-test: p = .97
sigtest(data=ex_survey1, cols = c_1, by = f_uni), # ANOVA/F-test: p = .22
sigtest(data=ex_survey1, cols = c_1, by = c_2)) # correlation: p = .976

sigtest(data=ex_survey1, cols = a_1, by = c_1) # NA
sigtest(data=ex_survey1, cols = b_1, by = c_1) # NA
}

}
