% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize a survey dataset for use in tables and graphs}
\usage{
summarize_data(
  data,
  cols = tidyselect::everything(),
  by = NULL,
  percentage = FALSE,
  showNA = c("ifany", "always", "no"),
  digits = 1,
  percent_sign = TRUE,
  sort_by = NULL,
  desc = FALSE,
  ignore_if_below = 0,
  label_separator = NULL,
  ...,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break output by.}

\item{percentage}{[\code{logical(1)}]\cr Whether to compute percentage or frequency (default).}

\item{showNA}{[\code{logical(1)}]\cr Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{percent_sign}{[\code{logical(1)}]\cr Whether to include percentage symbol on chart.}

\item{sort_by}{[\code{character(1)}] Sort output (and collapse if requested). Defaults to none (NULL).
\itemize{
\item{".top"}{The percentage for the highest category available in the variable.}
\item{".upper"}{The sum of the percentages for the categories above the middle category.}
\item{".mid_upper"}{The sum of the percentages for the categories including and above the middle category.}
\item{".mid_lower"}{The sum of the percentages for the categories including and below the middle category.}
\item{".lower"}{The sum of the percentages for the categories below the middle category.}
\item{".bottom"}{The percentage for the lowest category available in the variable.}
\item{".variable_label"}{Sort by the variable labels.}
\item{".id"}{Sort by the variable names.}
\item{".by_group"}{The groups of the by argument.}
\item{[\code{character(1)}]}{Character vector of category labels to sum together.}
}}

\item{desc}{[\code{logical(1)}]\cr Reverse sorting of sort_by. Defaults to ascending order (FALSE).}

\item{ignore_if_below}{[\code{numeric(1)}] Whether to hide label if below this value.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}

\item{...}{Optional parameters forwarded from above.}

\item{call}{Error call function, usually not needed.}
}
\value{
Dataset
}
\description{
Summarize a survey dataset for use in tables and graphs
}
\examples{
library(dplyr)
x <-
ex_survey1 \%>\%
saros:::summarize_data(cols = matches("^b_"),
                 sort_by = c("A lot", "A bit"),
                 label_separator = " - ", percentage = TRUE)
library(srvyr)
x <-
ex_survey1 \%>\%
  as_survey(strata = f_uni) \%>\%
  saros:::summarize_data(cols = matches("^b_"),
                                sort_by = c("A lot", "A bit"),
                                label_separator = " - ", percentage = TRUE)
}
