% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_ftp.R
\name{upload_to_ftp}
\alias{upload_to_ftp}
\title{Easily upload folder to ftp-server, with safe storage of credentials}
\usage{
upload_to_ftp(folder_path, ftp_path, ftp_server, username, password)
}
\arguments{
\item{folder_path}{String. Local path to upload.}

\item{ftp_path}{String. Remote path.}

\item{ftp_server}{String. Server, without "ftp://". Can be stored in .renviron for safe storage. Use SAROS_FTP_USERNAME}

\item{username}{String. Can be stored in .renviron for safe storage. Use SAROS_FTP_USERNAME}

\item{password}{String. Can be stored in .renviron for safe storage. Use SAROS_FTP_PASSWORD}
}
\value{
results from RCurl
}
\description{
Easily upload folder to ftp-server, with safe storage of credentials
}
