% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_character_vector_rule}
\alias{validate_character_vector_rule}
\title{Create character vector validation rule}
\usage{
validate_character_vector_rule(null_allowed = FALSE)
}
\arguments{
\item{null_allowed}{Whether NULL is an acceptable value (default FALSE)}
}
\description{
Create character vector validation rule
}
\keyword{internal}
