% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_double}
\alias{validate_double}
\title{Validate double/numeric parameter}
\usage{
validate_double(
  x,
  min = -Inf,
  max = Inf,
  null_allowed = FALSE,
  call = rlang::caller_env(),
  arg = rlang::caller_arg(x)
)
}
\description{
Validate double/numeric parameter
}
\keyword{internal}
