% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_double_rule}
\alias{validate_double_rule}
\title{Create double validation rule}
\usage{
validate_double_rule(min = -Inf, max = Inf, null_ok = FALSE)
}
\description{
Create double validation rule
}
\keyword{internal}
