% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_integerish}
\alias{validate_integerish}
\title{Validate integerish parameter}
\usage{
validate_integerish(
  x,
  min = -Inf,
  max = Inf,
  null_allowed = FALSE,
  call = rlang::caller_env(),
  arg = rlang::caller_arg(x)
)
}
\description{
Validate integerish parameter
}
\keyword{internal}
