% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_integerish_rule}
\alias{validate_integerish_rule}
\title{Create integerish validation rule}
\usage{
validate_integerish_rule(min = -Inf, max = Inf, null_ok = FALSE)
}
\description{
Create integerish validation rule
}
\keyword{internal}
