% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_string}
\alias{validate_string}
\title{Validate string parameter}
\usage{
validate_string(
  x,
  null_allowed = FALSE,
  call = rlang::caller_env(),
  arg = rlang::caller_arg(x)
)
}
\description{
Validate string parameter
}
\keyword{internal}
