% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_string_rule}
\alias{validate_string_rule}
\title{Create string validation rule}
\usage{
validate_string_rule(null_allowed = FALSE)
}
\arguments{
\item{null_allowed}{Whether NULL is an acceptable value (default FALSE)}
}
\description{
Create string validation rule
}
\keyword{internal}
